Consolidate 9 duplicate test files into tests/, config scanner   improvements
  - Merge 9 in-tree test files into canonical tests/ location, delete
  in-tree copies:
    test_configuration_scanner, test_credential_client,
  test_import_usage_analyzer,
    test_javascript_functions_parser, test_source_code_git_loader,
    test_transitive_detection, test_version_check,
  test_vulnerability_intel_sanitizer,
    test_web_patch_fetcher
  - Config scanner: add _CONFIG_DIR_ALLOWED_EXTENSIONS allowlist for
  directory-matched
    files, filtering out .js/.css/.map/.java/.py etc. from config
  collection
  - Config scanner: add build/tool config patterns (pyproject.toml,
  setup.cfg, tox.ini,
    tsconfig.json, .eslintrc.json, Makefile, CMakeLists.txt,
  meson.build)
  - Add MAVEN_OPTS with persistent repo cache to on-pull-request.yaml
  - Add 30+ new tests across merged files covering allowlist
  enforcement, build tool
    patterns, dependency manifests negative cases, and unique tests
  from in-tree files
