+++ .tekton/on-pull-request.yaml
@@ -368,6 +368,8 @@ spec:
                 # Pass the raw comment text into the container
                 - name: GOMODCACHE
                   value: "/exploit-iq-data/go/pkg/mod"
+                - name: MAVEN_OPTS
+                  value: "-Dmaven.repo.local=/exploit-iq-data/maven"
                 - name: UV_CACHE_DIR
                   value: "/tmp/uv-cache"
                 - name: UV_PYTHON_INSTALL_DIR
