--- src/exploit_iq_commons/utils/functions_parsers/tests/test_javascript_functions_parser.py
@@ -1,389 +1,0 @@
-import pytest
-from langchain_core.documents import Document
-
-from exploit_iq_commons.utils.functions_parsers.javascript_functions_parser import JavaScriptFunctionsParser
-
-
-@pytest.fixture
-def parser():
-    return JavaScriptFunctionsParser()
-
-
-class TestSearchForCalledFunction:
-
-    def test_this_method_call_same_class(self, parser):
-        callee_function = Document(
-            page_content="render() {\n    return '<div></div>';\n}//(class: Widget)",
-            metadata={"source": "widget.js", "content_type": "functions_classes"},
-        )
-        caller_function = Document(
-            page_content="update() {\n    this.render();\n}//(class: Widget)",
-            metadata={"source": "widget.js", "content_type": "functions_classes"},
-        )
-        code_documents = {
-            "widget.js": Document(
-                page_content=(
-                    "class Widget {\n"
-                    "    update() {\n"
-                    "        this.render();\n"
-                    "    }\n"
-                    "    render() {\n"
-                    "        return '<div></div>';\n"
-                    "    }\n"
-                    "}"
-                ),
-                metadata={"source": "widget.js"},
-            ),
-        }
-
-        result = parser.search_for_called_function(
-            caller_function, "render", callee_function, "", code_documents, [], "", {}, {}, []
-        )
-
-        assert result is True
-
-    def test_local_variable_type_match(self, parser):
-        callee_function = Document(
-            page_content="log(msg) {\n    console.log(msg);\n}//(class: Logger)",
-            metadata={"source": "logger.js", "content_type": "functions_classes"},
-        )
-        caller_function = Document(
-            page_content="function main() {\n    const logger = new Logger();\n    logger.log('hello');\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        code_documents = {
-            "app.js": Document(
-                page_content=(
-                    "const Logger = require('./logger');\n"
-                    "function main() {\n"
-                    "    const logger = new Logger();\n"
-                    "    logger.log('hello');\n"
-                    "}"
-                ),
-                metadata={"source": "app.js"},
-            ),
-            "logger.js": Document(
-                page_content="log(msg) {\n    console.log(msg);\n}//(class: Logger)",
-                metadata={"source": "logger.js"},
-            ),
-        }
-        functions_local_variables_index = {
-            "main@app.js": {"logger": {"type": "Logger"}},
-        }
-
-        result = parser.search_for_called_function(
-            caller_function, "log", callee_function, "logger-pkg", code_documents,
-            [], "logger.js", {}, functions_local_variables_index, []
-        )
-
-        assert result is True
-
-    def test_callee_not_called(self, parser):
-        """Caller body does not contain any call to the callee function."""
-        callee_function = Document(
-            page_content="doWork() {\n    return 42;\n}//(class: Worker)",
-            metadata={"source": "worker.js", "content_type": "functions_classes"},
-        )
-        caller_function = Document(
-            page_content="function main() {\n    console.log('hello');\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        code_documents = {
-            "app.js": Document(
-                page_content="function main() {\n    console.log('hello');\n}",
-                metadata={"source": "app.js"},
-            ),
-        }
-
-        result = parser.search_for_called_function(
-            caller_function, "doWork", callee_function, "worker-pkg", code_documents,
-            [], "worker.js", {}, {}, []
-        )
-
-        assert result is False
-
-
-class TestGetFunctionName:
-
-    def test_regular_function_declaration(self, parser):
-        func = Document(
-            page_content="function myFunc() { return 1; }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == "myFunc"
-
-    def test_arrow_function(self, parser):
-        func = Document(
-            page_content="const add = (a, b) => { return a + b; }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == "add"
-
-    def test_class_method_with_annotation(self, parser):
-        func = Document(
-            page_content="getValue() { return this.value; }//(class: MyClass)",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == "getValue"
-
-    def test_empty_content_returns_empty_string(self, parser):
-        """Empty page_content for a valid functions_classes document returns ''."""
-        func = Document(
-            page_content="",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == ""
-
-    def test_raises_value_error_for_class_declaration(self, parser):
-        func = Document(
-            page_content="class MyClass { constructor() {} }",
-            metadata={"content_type": "functions_classes"},
-        )
-        with pytest.raises(ValueError, match="Only function document is supported"):
-            parser.get_function_name(func)
-
-    def test_async_function(self, parser):
-        func = Document(
-            page_content="async function fetchData() { return await fetch('/api'); }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == "fetchData"
-
-    def test_export_default_anonymous_returns_empty(self, parser):
-        """export default function() should return empty string (anonymous)."""
-        func = Document(
-            page_content="export default function() { return 1; }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.get_function_name(func) == ""
-
-
-class TestGetFunctionCalls:
-
-    def test_direct_call(self, parser):
-        caller = Document(
-            page_content="function render() {\n    template();\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "template")
-        assert "template" in result
-
-    def test_qualified_call(self, parser):
-        caller = Document(
-            page_content="function render() {\n    lodash.template();\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "template")
-        assert "lodash.template" in result
-
-    def test_no_match(self, parser):
-        caller = Document(
-            page_content="function render() {\n    console.log('hi');\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "template")
-        assert result == []
-
-    def test_empty_callee_name(self, parser):
-        caller = Document(
-            page_content="function render() {\n    template();\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "")
-        assert result == []
-
-    def test_optional_chaining_call(self, parser):
-        """obj?.method() should match and strip the '?'."""
-        caller = Document(
-            page_content="function run() {\n    obj?.process();\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "process")
-        assert "obj.process" in result
-
-    def test_callback_reference(self, parser):
-        """Function passed as callback (no parens after name) is detected."""
-        caller = Document(
-            page_content="function run() {\n    setTimeout(cleanup, 1000);\n}",
-            metadata={"source": "app.js", "content_type": "functions_classes"},
-        )
-        result = parser._get_function_calls(caller, "cleanup")
-        assert "cleanup" in result
-
-
-class TestIsCommentLine:
-
-    def test_single_line_comment(self, parser):
-        assert parser.is_comment_line("// this is a comment") is True
-
-    def test_block_comment(self, parser):
-        assert parser.is_comment_line("/* block comment */") is True
-
-    def test_code_line(self, parser):
-        assert parser.is_comment_line("code();") is False
-
-    def test_indented_comment(self, parser):
-        assert parser.is_comment_line("   // indented comment") is True
-
-    def test_empty_line(self, parser):
-        assert parser.is_comment_line("") is False
-
-    def test_comment_after_code_is_not_comment_line(self, parser):
-        """A line starting with code is not a comment line even if it has a trailing comment."""
-        assert parser.is_comment_line("x = 1; // trailing") is False
-
-    def test_block_comment_continuation(self, parser):
-        """Block comment continuation line starting with *."""
-        assert parser.is_comment_line("   * @param foo") is True
-
-    def test_lone_asterisk(self, parser):
-        """Lone * is a comment continuation."""
-        assert parser.is_comment_line("   *") is True
-
-    def test_block_comment_end(self, parser):
-        """Block comment end */ is a comment line."""
-        assert parser.is_comment_line("   */") is True
-
-    def test_generator_method_not_comment(self, parser):
-        """Generator method *gen() must not be classified as a comment."""
-        assert parser.is_comment_line("  *myGenerator() {") is False
-
-    def test_generator_symbol_iterator_not_comment(self, parser):
-        """Generator *[Symbol.iterator]() must not be classified as a comment."""
-        assert parser.is_comment_line("  *[Symbol.iterator]() {") is False
-
-    def test_generator_with_space_not_comment(self, parser):
-        """Generator * gen() with space must not be classified as a comment."""
-        assert parser.is_comment_line("  * gen() {") is False
-
-
-class TestGetPackageNames:
-
-    def test_root_project_path(self, parser):
-        func = Document(
-            page_content="function util() {}",
-            metadata={"source": "src/utils.js", "content_type": "functions_classes"},
-        )
-        assert parser.get_package_names(func) == ["root_project"]
-
-    def test_third_party_package(self, parser):
-        func = Document(
-            page_content="function template() {}",
-            metadata={"source": "node_modules/lodash/index.js", "content_type": "functions_classes"},
-        )
-        assert parser.get_package_names(func) == ["lodash"]
-
-    def test_scoped_package(self, parser):
-        func = Document(
-            page_content="function transform() {}",
-            metadata={"source": "node_modules/@babel/core/index.js", "content_type": "functions_classes"},
-        )
-        assert parser.get_package_names(func) == ["@babel/core"]
-
-
-class TestTraceVariableToValue:
-
-    def test_string_double_quotes(self, parser):
-        lines = ['const foo = "hello";']
-        assert parser._trace_variable_to_value("foo", lines) == "hello"
-
-    def test_string_single_quotes(self, parser):
-        lines = ["const bar = 'world';"]
-        assert parser._trace_variable_to_value("bar", lines) == "world"
-
-    def test_variable_chain(self, parser):
-        """Tracing a variable assigned from another variable follows the chain."""
-        lines = [
-            'const a = "value";',
-            'const b = a;',
-        ]
-        assert parser._trace_variable_to_value("b", lines) == "value"
-
-    def test_variable_not_found(self, parser):
-        lines = ['const x = "something";']
-        assert parser._trace_variable_to_value("nothere", lines) == ""
-
-    def test_skips_comment_lines(self, parser):
-        """Comment lines containing a matching assignment are skipped."""
-        lines = [
-            '// const target = "wrong";',
-            'const target = "right";',
-        ]
-        assert parser._trace_variable_to_value("target", lines) == "right"
-
-    def test_circular_reference_returns_empty(self, parser):
-        """Circular variable references terminate without infinite recursion."""
-        lines = [
-            'const a = b;',
-            'const b = a;',
-        ]
-        assert parser._trace_variable_to_value("a", lines) == ""
-
-
-class TestIsFunction:
-
-    def test_function_content(self, parser):
-        doc = Document(
-            page_content="function hello() { return 1; }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.is_function(doc) is True
-
-    def test_class_content(self, parser):
-        doc = Document(
-            page_content="class MyClass { constructor() {} }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.is_function(doc) is False
-
-    def test_interface_content(self, parser):
-        doc = Document(
-            page_content="interface MyInterface { name: string; }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.is_function(doc) is False
-
-    def test_wrong_content_type(self, parser):
-        doc = Document(
-            page_content="function hello() { return 1; }",
-            metadata={"content_type": "full_source"},
-        )
-        assert parser.is_function(doc) is False
-
-    def test_exported_class(self, parser):
-        doc = Document(
-            page_content="export class Foo extends Bar { }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.is_function(doc) is False
-
-    def test_enum_content(self, parser):
-        doc = Document(
-            page_content="enum Direction { Up, Down }",
-            metadata={"content_type": "functions_classes"},
-        )
-        assert parser.is_function(doc) is False
-
-
-class TestIsPackageImported:
-
-    def test_es6_import_present(self, parser):
-        code = "import { template } from 'lodash';\nfunction render() { template(); }"
-        assert parser.is_package_imported(code, "template", "lodash") is True
-
-    def test_require_import_present(self, parser):
-        code = "const template = require('lodash');\nfunction render() { template(); }"
-        assert parser.is_package_imported(code, "template", "lodash") is True
-
-    def test_no_import(self, parser):
-        code = "function render() { template(); }"
-        assert parser.is_package_imported(code, "template", "lodash") is False
-
-    def test_empty_identifier(self, parser):
-        code = "import { template } from 'lodash';"
-        assert parser.is_package_imported(code, "", "lodash") is False
-
-    def test_import_star_as_named_identifier(self, parser):
-        """import * as alias from 'pkg' — the alias is a distinct identifier from the package."""
-        code = "import * as _ from 'lodash';\nfunction render() { _.template(); }"
-        assert parser.is_package_imported(code, "_", "lodash") is True
