--- src/vuln_analysis/tools/tests/test_configuration_scanner.py
@@ -1,386 +1,0 @@
-import os
-import tempfile
-
-import pytest
-
-from vuln_analysis.tools.configuration_scanner import (
-    _BINARY_EXTENSIONS,
-    _CONFIG_DIR_ALLOWED_EXTENSIONS,
-    _collect_config_files,
-    _count_config_matches,
-    _is_config_file,
-    format_context_snippet,
-    search_config_content,
-)
-
-
-class TestFormatContextSnippet:
-    def test_match_at_first_line_with_context(self):
-        lines = ["alpha", "bravo", "charlie", "delta", "echo"]
-        result = format_context_snippet(lines, match_line=0, context_lines=2)
-        output_lines = result.split("\n")
-        assert output_lines[0] == "> 1: alpha"
-        assert output_lines[1] == "  2: bravo"
-        assert output_lines[2] == "  3: charlie"
-        assert len(output_lines) == 3
-
-    def test_match_at_last_line_with_context(self):
-        lines = ["alpha", "bravo", "charlie", "delta", "echo"]
-        result = format_context_snippet(lines, match_line=4, context_lines=2)
-        output_lines = result.split("\n")
-        assert output_lines[0] == "  3: charlie"
-        assert output_lines[1] == "  4: delta"
-        assert output_lines[2] == "> 5: echo"
-        assert len(output_lines) == 3
-
-    def test_zero_context_lines(self):
-        lines = ["alpha", "bravo", "charlie"]
-        result = format_context_snippet(lines, match_line=1, context_lines=0)
-        output_lines = result.split("\n")
-        assert output_lines == ["> 2: bravo"]
-
-    def test_middle_match_with_full_context(self):
-        lines = ["one", "two", "three", "four", "five"]
-        result = format_context_snippet(lines, match_line=2, context_lines=1)
-        output_lines = result.split("\n")
-        assert output_lines[0] == "  2: two"
-        assert output_lines[1] == "> 3: three"
-        assert output_lines[2] == "  4: four"
-        assert len(output_lines) == 3
-
-    def test_single_line_file(self):
-        lines = ["only"]
-        result = format_context_snippet(lines, match_line=0, context_lines=5)
-        assert result == "> 1: only"
-
-
-class TestCountConfigMatches:
-    def test_counts_matching_lines(self):
-        config_files = [
-            ("app.properties", "server.port=8080\nssl.enabled=true\nlogging.level=INFO"),
-            ("db.yml", "host: localhost\nport: 5432\nssl: false"),
-        ]
-        keywords = ["ssl"]
-        assert _count_config_matches(config_files, keywords) == 2
-
-    def test_case_insensitive_matching(self):
-        config_files = [
-            ("app.yml", "SSL_ENABLED=true\nSsl_Mode=require"),
-        ]
-        keywords = ["ssl"]
-        assert _count_config_matches(config_files, keywords) == 2
-
-    def test_no_matches(self):
-        config_files = [
-            ("app.yml", "host: localhost\nport: 8080"),
-        ]
-        keywords = ["ssl", "tls"]
-        assert _count_config_matches(config_files, keywords) == 0
-
-    def test_multiple_keywords_same_line(self):
-        config_files = [
-            ("app.conf", "ssl_port=443"),
-        ]
-        keywords = ["ssl", "port"]
-        assert _count_config_matches(config_files, keywords) == 1
-
-    def test_empty_config_files(self):
-        assert _count_config_matches([], ["ssl"]) == 0
-
-
-class TestCollectConfigFiles:
-    def test_skips_binary_extension_files(self):
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "config")
-            os.makedirs(conf_dir)
-
-            text_cfg = os.path.join(conf_dir, "settings.yml")
-            with open(text_cfg, "w") as f:
-                f.write("key: value")
-
-            binary_png = os.path.join(conf_dir, "logo.png")
-            with open(binary_png, "wb") as f:
-                f.write(b"\x89PNG fake")
-
-            binary_jar = os.path.join(conf_dir, "lib.jar")
-            with open(binary_jar, "wb") as f:
-                f.write(b"PK fake jar")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(path) for path, _ in results]
-
-            assert "settings.yml" in collected_names
-            assert "logo.png" not in collected_names
-            assert "lib.jar" not in collected_names
-
-    def test_collects_files_in_config_dir(self):
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "resources")
-            os.makedirs(conf_dir)
-
-            app_cfg = os.path.join(conf_dir, "database.xml")
-            with open(app_cfg, "w") as f:
-                f.write("<database><host>localhost</host></database>")
-
-            results = _collect_config_files(repo)
-            collected_paths = [path for path, _ in results]
-
-            assert any("database.xml" in p for p in collected_paths)
-
-    def test_collects_config_pattern_files(self):
-        with tempfile.TemporaryDirectory() as repo:
-            props = os.path.join(repo, "application.properties")
-            with open(props, "w") as f:
-                f.write("server.port=8080")
-
-            env_file = os.path.join(repo, "app.env")
-            with open(env_file, "w") as f:
-                f.write("DB_HOST=localhost")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(path) for path, _ in results]
-
-            assert "application.properties" in collected_names
-            assert "app.env" in collected_names
-
-    def test_skips_git_and_pycache_dirs(self):
-        with tempfile.TemporaryDirectory() as repo:
-            git_conf = os.path.join(repo, ".git", "config")
-            os.makedirs(os.path.dirname(git_conf))
-            with open(git_conf, "w") as f:
-                f.write("[core]")
-
-            pycache_conf = os.path.join(repo, "__pycache__", "config.yml")
-            os.makedirs(os.path.dirname(pycache_conf))
-            with open(pycache_conf, "w") as f:
-                f.write("cached: true")
-
-            results = _collect_config_files(repo)
-            collected_paths = [path for path, _ in results]
-
-            assert not any(p.startswith(".git/") or p == ".git" for p in collected_paths)
-            assert not any("__pycache__" in p for p in collected_paths)
-
-    def test_skips_files_over_500kb(self):
-        with tempfile.TemporaryDirectory() as repo:
-            large_cfg = os.path.join(repo, "application.properties")
-            with open(large_cfg, "w") as f:
-                f.write("x" * 600_000)
-
-            small_cfg = os.path.join(repo, "config.yml")
-            with open(small_cfg, "w") as f:
-                f.write("small: true")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(path) for path, _ in results]
-
-            assert "config.yml" in collected_names
-            assert "application.properties" not in collected_names
-
-    def test_returns_file_content(self):
-        with tempfile.TemporaryDirectory() as repo:
-            cfg = os.path.join(repo, "application.yml")
-            with open(cfg, "w") as f:
-                f.write("server:\n  port: 9090")
-
-            results = _collect_config_files(repo)
-            assert len(results) == 1
-            _, content = results[0]
-            assert "port: 9090" in content
-
-
-class TestBinaryExtensions:
-    def test_contains_expected_extensions(self):
-        expected = {".jar", ".png", ".jpg", ".pdf", ".class", ".zip", ".pyc"}
-        assert expected.issubset(_BINARY_EXTENSIONS)
-
-    def test_is_frozenset(self):
-        assert isinstance(_BINARY_EXTENSIONS, frozenset)
-
-
-class TestSearchConfigContent:
-    def test_returns_formatted_matches(self):
-        config_files = [
-            ("app.properties", "server.port=8080\nssl.enabled=true\nlogging.level=INFO"),
-        ]
-        result = search_config_content(config_files, ["ssl"], context_lines=0)
-        assert len(result) == 1
-        assert "app.properties" in result[0]
-        assert "ssl.enabled=true" in result[0]
-
-    def test_respects_max_results(self):
-        config_files = [("big.yml", "\n".join(f"keyword_{i}" for i in range(50)))]
-        result = search_config_content(config_files, ["keyword"], max_results=3, context_lines=0)
-        assert len(result) == 3
-
-    def test_no_matches_returns_empty(self):
-        config_files = [("app.yml", "host: localhost")]
-        result = search_config_content(config_files, ["nonexistent"])
-        assert result == []
-
-    def test_case_insensitive(self):
-        config_files = [("app.yml", "SSL_ENABLED=true")]
-        result = search_config_content(config_files, ["ssl"], context_lines=0)
-        assert len(result) == 1
-
-    def test_includes_source_label(self):
-        config_files = [("app.yml", "match_keyword")]
-        result = search_config_content(config_files, ["match"], source_label="git://repo", context_lines=0)
-        assert "git://repo" in result[0]
-
-
-class TestWildcardPatterns:
-    def test_dockerfile_variants(self):
-        assert _is_config_file("Dockerfile") is True
-        assert _is_config_file("Dockerfile.prod") is True
-        assert _is_config_file("Dockerfile.dev") is True
-
-    def test_docker_compose_variants(self):
-        assert _is_config_file("docker-compose.yml") is True
-        assert _is_config_file("docker-compose.prod.yml") is True
-        assert _is_config_file("docker-compose-dev.yaml") is True
-
-    def test_non_config_with_config_extension(self):
-        """Files with config extensions (.properties, .env, .conf, .ini) match anywhere."""
-        assert _is_config_file("random.properties") is True
-        assert _is_config_file("random.env") is True
-        assert _is_config_file("random.conf") is True
-        assert _is_config_file("random.ini") is True
-
-    def test_build_tool_config_files(self):
-        """Build/tool configuration files matched by exact name."""
-        assert _is_config_file("pyproject.toml") is True
-        assert _is_config_file("setup.cfg") is True
-        assert _is_config_file("tox.ini") is True
-        assert _is_config_file("tsconfig.json") is True
-        assert _is_config_file(".eslintrc.json") is True
-        assert _is_config_file("Makefile") is True
-        assert _is_config_file("CMakeLists.txt") is True
-        assert _is_config_file("meson.build") is True
-
-    def test_dependency_manifests_not_matched(self):
-        """Dependency management files should not be matched as config."""
-        assert _is_config_file("go.mod") is False
-        assert _is_config_file("go.sum") is False
-        assert _is_config_file("requirements.txt") is False
-        assert _is_config_file("package.json") is False
-        assert _is_config_file("Cargo.toml") is False
-        assert _is_config_file("Gemfile") is False
-        assert _is_config_file("pom.xml") is False
-
-    def test_source_code_not_matched(self):
-        assert _is_config_file("main.java") is False
-        assert _is_config_file("app.js") is False
-        assert _is_config_file("util.py") is False
-        assert _is_config_file("handler.go") is False
-
-
-class TestConfigDirAllowedExtensions:
-    """Only files with config-like extensions should be collected from config directories."""
-
-    def test_js_in_resources_dir_skipped(self):
-        """Reproduces Keycloak issue: .js files under resources/ were collected as configs."""
-        with tempfile.TemporaryDirectory() as repo:
-            res_dir = os.path.join(repo, "src", "main", "resources", "theme")
-            os.makedirs(res_dir)
-
-            for fname in ["react-core.js", "react-core.js.map", "app.css"]:
-                with open(os.path.join(res_dir, fname), "w") as f:
-                    f.write("not a config")
-
-            yml_cfg = os.path.join(res_dir, "settings.yml")
-            with open(yml_cfg, "w") as f:
-                f.write("key: value")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-
-            assert "settings.yml" in collected_names
-            assert "react-core.js" not in collected_names
-            assert "react-core.js.map" not in collected_names
-            assert "app.css" not in collected_names
-
-    def test_java_source_in_resources_dir_skipped(self):
-        with tempfile.TemporaryDirectory() as repo:
-            res_dir = os.path.join(repo, "resources")
-            os.makedirs(res_dir)
-
-            with open(os.path.join(res_dir, "Helper.java"), "w") as f:
-                f.write("public class Helper {}")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-            assert "Helper.java" not in collected_names
-
-    def test_named_config_in_resources_dir_still_collected(self):
-        """application.properties matched by name pattern bypasses the extension filter."""
-        with tempfile.TemporaryDirectory() as repo:
-            res_dir = os.path.join(repo, "resources")
-            os.makedirs(res_dir)
-
-            with open(os.path.join(res_dir, "application.properties"), "w") as f:
-                f.write("server.port=8080")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-            assert "application.properties" in collected_names
-
-    def test_xml_in_config_dir_collected(self):
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "config")
-            os.makedirs(conf_dir)
-
-            with open(os.path.join(conf_dir, "persistence.xml"), "w") as f:
-                f.write("<persistence/>")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-            assert "persistence.xml" in collected_names
-
-    def test_json_in_config_dir_collected(self):
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "etc")
-            os.makedirs(conf_dir)
-
-            with open(os.path.join(conf_dir, "config.json"), "w") as f:
-                f.write('{"key": "value"}')
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-            assert "config.json" in collected_names
-
-    def test_extensionless_file_in_config_dir_collected(self):
-        """Extensionless files like Caddyfile or Makefile should pass the filter."""
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "config")
-            os.makedirs(conf_dir)
-
-            with open(os.path.join(conf_dir, "Caddyfile"), "w") as f:
-                f.write(":8080 { respond 200 }")
-
-            results = _collect_config_files(repo)
-            collected_names = [os.path.basename(p) for p, _ in results]
-            assert "Caddyfile" in collected_names
-
-    def test_all_allowed_extensions_collected(self):
-        """Every extension in the allowlist should be collected from a config dir."""
-        with tempfile.TemporaryDirectory() as repo:
-            conf_dir = os.path.join(repo, "conf")
-            os.makedirs(conf_dir)
-
-            for ext in _CONFIG_DIR_ALLOWED_EXTENSIONS:
-                fname = f"test{ext}"
-                with open(os.path.join(conf_dir, fname), "w") as f:
-                    f.write("content")
-
-            results = _collect_config_files(repo)
-            collected_names = {os.path.basename(p) for p, _ in results}
-            for ext in _CONFIG_DIR_ALLOWED_EXTENSIONS:
-                assert f"test{ext}" in collected_names, f"Expected test{ext} to be collected"
-
-    def test_allowlist_is_frozenset(self):
-        assert isinstance(_CONFIG_DIR_ALLOWED_EXTENSIONS, frozenset)
-
-    def test_contains_core_config_extensions(self):
-        expected = {".xml", ".yml", ".yaml", ".json", ".toml", ".properties", ".conf", ".ini", ".env"}
-        assert expected.issubset(_CONFIG_DIR_ALLOWED_EXTENSIONS)
