--- src/vuln_analysis/tools/tests/test_credential_client.py
@@ -1,388 +1,0 @@
-"""Tests for fetch_and_decrypt_credential public API.
-
-Covers: successful PAT and SSH key decryption, URL construction,
-auth header correctness, HTTP error codes, decryption errors,
-network errors, and secret-not-in-logs verification.
-"""
-
-import base64
-import logging
-from unittest.mock import MagicMock, patch
-
-import pytest
-from cryptography.hazmat.primitives.ciphers.aead import AESGCM
-
-from exploit_iq_commons.utils.credential_client import (
-    AuthenticationError,
-    CredentialNotFoundError,
-    DecryptionError,
-    TLSConfigurationError,
-    _credential_id_ctx,
-    _validate_ca_bundle,
-    credential_context,
-    fetch_and_decrypt_credential,
-)
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-_ENCRYPTION_KEY = "test-encryption-key-32-bytes!!!!"  # exactly 32 UTF-8 bytes
-_IV = b"\x00" * 12  # deterministic IV for tests
-
-
-def _encrypt(plaintext: str, key: str = _ENCRYPTION_KEY, iv: bytes = _IV) -> bytes:
-    """Encrypt plaintext using AES-256-GCM with the same parameters as the backend."""
-    aesgcm = AESGCM(key.encode("utf-8")[:32])
-    return aesgcm.encrypt(iv, plaintext.encode("utf-8"), None)
-
-
-def _make_response(
-    plaintext: str,
-    credential_type: str = "PAT",
-    username: str | None = "github-bot",
-    user_id: str = "alice@example.com",
-    key: str = _ENCRYPTION_KEY,
-    iv: bytes = _IV,
-) -> dict:
-    encrypted = _encrypt(plaintext, key=key, iv=iv)
-    return {
-        "encryptedSecretValue": base64.b64encode(encrypted).decode(),
-        "iv": base64.b64encode(iv).decode(),
-        "username": username,
-        "credentialType": credential_type,
-        "userId": user_id,
-    }
-
-
-def _mock_http_ok(payload: dict) -> MagicMock:
-    resp = MagicMock()
-    resp.status_code = 200
-    resp.ok = True
-    resp.json.return_value = payload
-    return resp
-
-
-def _mock_http_error(status: int) -> MagicMock:
-    resp = MagicMock()
-    resp.status_code = status
-    resp.ok = False
-    return resp
-
-
-# ---------------------------------------------------------------------------
-# Tests: successful decryption
-# ---------------------------------------------------------------------------
-
-class TestFetchAndDecryptCredential:
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_pat_decryption_success(self, mock_get, mock_ca):
-        plaintext = "ghp_myPersonalAccessToken123"
-        mock_get.return_value = _mock_http_ok(_make_response(plaintext, credential_type="PAT"))
-
-        result = fetch_and_decrypt_credential(
-            credential_id="cred-uuid-1",
-            jwt_token="scan.jwt.token",
-            backend_url="https://backend.example.com",
-            encryption_key=_ENCRYPTION_KEY,
-        )
-
-        assert result["secret_value"] == plaintext
-        assert result["username"] == "github-bot"
-        assert result["credential_type"] == "PAT"
-        assert result["user_id"] == "alice@example.com"
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_ssh_key_decryption_success(self, mock_get, mock_ca):
-        plaintext = "-----BEGIN OPENSSH PRIVATE KEY-----\nabc123\n-----END OPENSSH PRIVATE KEY-----"
-        mock_get.return_value = _mock_http_ok(
-            _make_response(plaintext, credential_type="SSH_KEY", username=None)
-        )
-
-        result = fetch_and_decrypt_credential(
-            credential_id="cred-uuid-2",
-            jwt_token="scan.jwt.token",
-            backend_url="https://backend.example.com",
-            encryption_key=_ENCRYPTION_KEY,
-        )
-
-        assert result["secret_value"] == plaintext
-        assert result["username"] is None
-        assert result["credential_type"] == "SSH_KEY"
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_correct_url_and_auth_header(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_ok(_make_response("token"))
-
-        fetch_and_decrypt_credential(
-            credential_id="cred-uuid-3",
-            jwt_token="my.jwt",
-            backend_url="https://backend.example.com",
-            encryption_key=_ENCRYPTION_KEY,
-        )
-
-        mock_get.assert_called_once_with(
-            "https://backend.example.com/api/v1/credentials/cred-uuid-3",
-            headers={"Authorization": "Bearer my.jwt"},
-            timeout=10,
-            verify=False,
-        )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value="/tmp/test-ca.crt")
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_tls_ca_bundle_applied_when_valid(self, mock_get, mock_ca):
-        """When _validate_ca_bundle returns a path, requests.get uses it for TLS verification."""
-        plaintext = "ghp_tlsVerifiedToken"
-        mock_get.return_value = _mock_http_ok(_make_response(plaintext))
-
-        result = fetch_and_decrypt_credential(
-            credential_id="cred-tls",
-            jwt_token="scan.jwt.token",
-            backend_url="https://backend.example.com",
-            encryption_key=_ENCRYPTION_KEY,
-        )
-
-        assert result["secret_value"] == plaintext
-        mock_get.assert_called_once_with(
-            "https://backend.example.com/api/v1/credentials/cred-tls",
-            headers={"Authorization": "Bearer scan.jwt.token"},
-            timeout=10,
-            verify="/tmp/test-ca.crt",
-        )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_backend_url_trailing_slash_stripped(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_ok(_make_response("token"))
-
-        fetch_and_decrypt_credential(
-            credential_id="cred-uuid-4",
-            jwt_token="my.jwt",
-            backend_url="https://backend.example.com/",  # trailing slash
-            encryption_key=_ENCRYPTION_KEY,
-        )
-
-        call_url = mock_get.call_args[0][0]
-        assert "//" not in call_url.split("://")[1], "Double slash in URL"
-
-
-# ---------------------------------------------------------------------------
-# Tests: error handling
-# ---------------------------------------------------------------------------
-
-class TestErrorHandling:
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_404_raises_credential_not_found(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_error(404)
-
-        with pytest.raises(CredentialNotFoundError):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_401_raises_authentication_error(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_error(401)
-
-        with pytest.raises(AuthenticationError):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_403_raises_authentication_error(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_error(403)
-
-        with pytest.raises(AuthenticationError):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_500_raises_runtime_error(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_error(500)
-
-        with pytest.raises(RuntimeError):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_wrong_key_raises_decryption_error(self, mock_get, mock_ca):
-        mock_get.return_value = _mock_http_ok(_make_response("secret", key=_ENCRYPTION_KEY))
-
-        with pytest.raises(DecryptionError):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key="wrong-key-32-bytes-xxxxxxxxxxxxxx",
-            )
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_network_error_raises_runtime_error(self, mock_get, mock_ca):
-        import requests as req_lib
-        mock_get.side_effect = req_lib.RequestException("Connection refused")
-
-        with pytest.raises(RuntimeError, match="Network error"):
-            fetch_and_decrypt_credential(
-                credential_id="cred-id",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-
-# ---------------------------------------------------------------------------
-# Tests: secret_value never appears in logs
-# ---------------------------------------------------------------------------
-
-class TestNoPlaintextInLogs:
-
-    @patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle", return_value=False)
-    @patch("exploit_iq_commons.utils.credential_client.requests.get")
-    def test_secret_not_logged(self, mock_get, mock_ca, caplog):
-        secret = "super-secret-token-must-not-appear-in-logs"
-        mock_get.return_value = _mock_http_ok(_make_response(secret))
-
-        with caplog.at_level(logging.DEBUG, logger="exploit_iq_commons.utils.credential_client"):
-            fetch_and_decrypt_credential(
-                credential_id="cred-uuid-log",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-
-        for record in caplog.records:
-            assert secret not in record.getMessage(), (
-                f"Secret value leaked into log message: {record.getMessage()}"
-            )
-
-
-# ---------------------------------------------------------------------------
-# Tests: TLS / HTTP URL path selection
-# ---------------------------------------------------------------------------
-
-class TestTLSPathSelection:
-
-    def test_http_url_skips_tls_validation(self):
-        """HTTP URLs skip CA bundle validation entirely."""
-        mock_get = MagicMock(return_value=_mock_http_ok(_make_response("token")))
-        with patch("exploit_iq_commons.utils.credential_client.requests.get", mock_get), \
-             patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle") as mock_ca:
-            result = fetch_and_decrypt_credential(
-                credential_id="cred-http",
-                jwt_token="jwt",
-                backend_url="http://localhost:8080",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-        assert result["secret_value"] == "token"
-        mock_get.assert_called_once()
-        assert mock_get.call_args[1]["verify"] is False
-        mock_ca.assert_not_called()
-
-    def test_custom_ca_bundle_from_env(self, monkeypatch):
-        """CLIENT_CA_BUNDLE env var overrides the default CA bundle path."""
-        monkeypatch.setenv("CLIENT_CA_BUNDLE", "/custom/ca.crt")
-        with patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle",
-                    return_value="/custom/ca.crt") as mock_ca, \
-             patch("exploit_iq_commons.utils.credential_client.requests.get") as mock_get:
-            mock_get.return_value = _mock_http_ok(_make_response("token"))
-            fetch_and_decrypt_credential(
-                credential_id="cred-env",
-                jwt_token="jwt",
-                backend_url="https://backend.example.com",
-                encryption_key=_ENCRYPTION_KEY,
-            )
-        mock_ca.assert_called_once_with("/custom/ca.crt")
-
-    def test_tls_configuration_error_propagates(self):
-        """TLSConfigurationError from CA bundle validation propagates uncaught."""
-        with patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle",
-                    side_effect=TLSConfigurationError("CA bundle not found")):
-            with pytest.raises(TLSConfigurationError, match="CA bundle not found"):
-                fetch_and_decrypt_credential(
-                    credential_id="cred-tls",
-                    jwt_token="jwt",
-                    backend_url="https://backend.example.com",
-                    encryption_key=_ENCRYPTION_KEY,
-                )
-
-
-# ---------------------------------------------------------------------------
-# Tests: decryption edge cases
-# ---------------------------------------------------------------------------
-
-class TestDecryptionEdgeCases:
-
-    def test_unexpected_decryption_error_general_exception(self):
-        """Generic exceptions during AES decryption raise DecryptionError."""
-        with patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle",
-                    return_value="/tmp/test-ca.crt"), \
-             patch("exploit_iq_commons.utils.credential_client.requests.get") as mock_get, \
-             patch("exploit_iq_commons.utils.credential_client.AESGCM") as mock_aesgcm:
-            mock_get.return_value = _mock_http_ok(_make_response("token"))
-            mock_aesgcm.return_value.decrypt.side_effect = TypeError("unexpected")
-            with pytest.raises(DecryptionError, match="unexpected general failure"):
-                fetch_and_decrypt_credential(
-                    credential_id="cred-general",
-                    jwt_token="jwt",
-                    backend_url="https://backend.example.com",
-                    encryption_key=_ENCRYPTION_KEY,
-                )
-
-    def test_missing_encrypted_fields_raises_decryption_error(self):
-        """Missing encryptedSecretValue/iv fields raise DecryptionError."""
-        bad_payload = {"credentialType": "PAT"}  # missing encryptedSecretValue and iv
-        with patch("exploit_iq_commons.utils.credential_client._validate_ca_bundle",
-                    return_value="/tmp/test-ca.crt"), \
-             patch("exploit_iq_commons.utils.credential_client.requests.get") as mock_get:
-            resp = MagicMock()
-            resp.status_code = 200
-            resp.ok = True
-            resp.json.return_value = bad_payload
-            mock_get.return_value = resp
-            with pytest.raises(DecryptionError, match="Invalid response payload"):
-                fetch_and_decrypt_credential(
-                    credential_id="cred-missing",
-                    jwt_token="jwt",
-                    backend_url="https://backend.example.com",
-                    encryption_key=_ENCRYPTION_KEY,
-                )
-
-
-# ---------------------------------------------------------------------------
-# Tests: credential_context
-# ---------------------------------------------------------------------------
-
-class TestCredentialContext:
-
-    def test_none_credential_id(self):
-        """credential_context(None) sets ctx to None and makes no HTTP calls."""
-        with patch("exploit_iq_commons.utils.credential_client.requests.get") as mock_get:
-            with credential_context(None):
-                assert _credential_id_ctx.get() is None
-            mock_get.assert_not_called()
