--- src/vuln_analysis/tools/tests/test_import_usage_analyzer.py
@@ -1,306 +1,0 @@
-import re
-from unittest.mock import MagicMock, patch, PropertyMock
-
-import pytest
-
-from vuln_analysis.tools.import_usage_analyzer import analyze_imports
-from exploit_iq_commons.utils.functions_parsers.lang_functions_parsers import LanguageFunctionsParser
-
-
-def _make_doc(file_path: str, content: str) -> dict:
-    return {"file_path": [file_path], "content": [content]}
-
-
-def _make_searcher(docs: list[dict], error_indices: set[int] | None = None):
-    error_indices = error_indices or set()
-    hits = [(1.0, i) for i in range(len(docs))]
-
-    searcher = MagicMock()
-    type(searcher).num_docs = PropertyMock(return_value=len(docs))
-
-    search_result = MagicMock()
-    type(search_result).hits = PropertyMock(return_value=hits)
-    searcher.search.return_value = search_result
-
-    def doc_side_effect(doc_address):
-        if doc_address in error_indices:
-            raise ValueError(f"corrupted doc at {doc_address}")
-        return docs[doc_address]
-
-    searcher.doc.side_effect = doc_side_effect
-
-    # Store expected query sentinel so tests can verify the correct query is passed
-    searcher._expected_query = "ALL"
-    return searcher
-
-
-@patch("vuln_analysis.tools.import_usage_analyzer.tantivy")
-def test_analyze_imports_skips_errored_doc(mock_tantivy):
-    mock_tantivy.Query.all_query.return_value = "ALL"
-    docs = [
-        _make_doc("src/app/Main.java", "import com.example.lib;\nlib.doSomething();"),
-        _make_doc("src/app/Other.java", "import com.example.lib;\nlib.callMethod();"),
-    ]
-    searcher = _make_searcher(docs, error_indices={0})
-    patterns = [re.compile(re.escape("com.example.lib"), re.IGNORECASE)]
-
-    result = analyze_imports(searcher, patterns, "com.example.lib")
-
-    assert "Other.java" in result
-    assert "Main.java" not in result
-    searcher.search.assert_called_once_with("ALL", limit=2)
-
-
-@patch("vuln_analysis.tools.import_usage_analyzer.tantivy")
-def test_analyze_imports_trims_to_max_files(mock_tantivy):
-    mock_tantivy.Query.all_query.return_value = "ALL"
-    docs = [
-        _make_doc(f"src/app/File{i}.java", f"import com.example.lib;\nlib.use{i}();")
-        for i in range(25)
-    ]
-    searcher = _make_searcher(docs)
-    patterns = [re.compile(re.escape("com.example.lib"), re.IGNORECASE)]
-
-    result = analyze_imports(searcher, patterns, "com.example.lib", max_files=20)
-
-    matched_files = [f"File{i}.java" for i in range(25) if f"File{i}.java" in result]
-    assert len(matched_files) <= 20
-    assert "20 of 25 results" in result
-
-
-@patch("vuln_analysis.tools.import_usage_analyzer.tantivy")
-def test_analyze_imports_continues_after_multiple_errors(mock_tantivy):
-    """Multiple errored docs should be skipped without affecting valid results."""
-    mock_tantivy.Query.all_query.return_value = "ALL"
-    docs = [
-        _make_doc("src/app/Bad1.java", "content"),
-        _make_doc("src/app/Bad2.java", "content"),
-        _make_doc("src/app/Good.java", "import com.example.lib;\nlib.doSomething();"),
-    ]
-    searcher = _make_searcher(docs, error_indices={0, 1})
-    patterns = [re.compile(re.escape("com.example.lib"), re.IGNORECASE)]
-
-    result = analyze_imports(searcher, patterns, "com.example.lib")
-
-    assert "Good.java" in result
-    assert "Bad1.java" not in result
-    assert "Bad2.java" not in result
-
-
-@patch("vuln_analysis.tools.import_usage_analyzer.tantivy")
-def test_analyze_imports_default_max_files(mock_tantivy):
-    """Default max_files=20 should limit output even with more matching docs."""
-    mock_tantivy.Query.all_query.return_value = "ALL"
-    docs = [
-        _make_doc(f"src/app/File{i}.java", f"import com.example.lib;\nlib.call{i}();")
-        for i in range(30)
-    ]
-    searcher = _make_searcher(docs)
-    patterns = [re.compile(re.escape("com.example.lib"), re.IGNORECASE)]
-
-    result = analyze_imports(searcher, patterns, "com.example.lib")
-
-    # Default max_files=20, so with 30 matching docs only 20 appear in output
-    assert "20 of 30 results" in result
-
-
-def test_get_import_search_patterns_empty_string():
-    patterns = LanguageFunctionsParser.get_import_search_patterns(None, "")
-
-    assert isinstance(patterns, list)
-    assert len(patterns) == 1
-    assert isinstance(patterns[0], re.Pattern)
-    assert patterns[0].search("anything") is not None
-
-
-# ---------------------------------------------------------------------------
-# Helpers for max_files trimming tests
-# ---------------------------------------------------------------------------
-
-def _build_mock_searcher(file_entries: list[tuple[str, str]]):
-    """Build a mock tantivy searcher returning the given (file_path, content) pairs."""
-    searcher = MagicMock()
-    searcher.num_docs = len(file_entries)
-
-    hits = [(1.0, idx) for idx in range(len(file_entries))]
-    search_result = MagicMock()
-    search_result.hits = hits
-    searcher.search.return_value = search_result
-
-    doc_map = {}
-    for idx, (fp, content) in enumerate(file_entries):
-        doc_map[idx] = {"file_path": [fp], "content": [content]}
-    searcher.doc.side_effect = lambda addr: doc_map[addr]
-
-    return searcher
-
-
-def _make_import_pattern(package: str) -> list[re.Pattern]:
-    """Create a simple import pattern that matches 'import <package>'."""
-    return [re.compile(rf'import\s+{re.escape(package)}', re.IGNORECASE)]
-
-
-class TestMaxFilesTrimming:
-    """Test that app results are prioritised over dep results when
-    total exceeds max_files."""
-
-    def test_app_results_prioritised_over_dep(self):
-        """When total results exceed max_files, app results fill the budget
-        first, then remaining goes to dep results."""
-        package = "com.example.lib"
-        entries = [
-            ("src/main/java/App1.java", f"import {package};\nApp1.use();"),
-            ("src/main/java/App2.java", f"import {package};\nApp2.use();"),
-            ("src/main/java/App3.java", f"import {package};\nApp3.use();"),
-            ("dependencies-sources/lib-1.0-sources/Dep1.java", f"import {package};\nDep1.use();"),
-            ("dependencies-sources/lib-1.0-sources/Dep2.java", f"import {package};\nDep2.use();"),
-            ("dependencies-sources/lib-1.0-sources/Dep3.java", f"import {package};\nDep3.use();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=4)
-
-        assert "Main application (3 of 3 results)" in result
-        assert "Application library dependencies (1 of 3 results)" in result
-
-    def test_app_fills_budget_no_dep(self):
-        """When app results >= max_files, dep results get 0 budget."""
-        package = "com.example.lib"
-        entries = [
-            ("src/main/java/App1.java", f"import {package};\nuse();"),
-            ("src/main/java/App2.java", f"import {package};\nuse();"),
-            ("src/main/java/App3.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep1.java", f"import {package};\nuse();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=2)
-
-        assert "Main application (2 of 3 results)" in result
-        assert "Application library dependencies (0 of 1 results)" in result
-
-    def test_app_under_budget_remaining_to_dep(self):
-        """When app results < max_files, remaining budget goes to dep."""
-        package = "com.example.lib"
-        entries = [
-            ("src/main/java/App1.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep1.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep2.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep3.java", f"import {package};\nuse();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=3)
-
-        assert "Main application (1 of 1 results)" in result
-        assert "Application library dependencies (2 of 3 results)" in result
-
-    def test_all_results_fit_within_budget(self):
-        """When total results <= max_files, all are shown."""
-        package = "com.example.lib"
-        entries = [
-            ("src/main/java/App1.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep1.java", f"import {package};\nuse();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=10)
-
-        assert "Main application (1 of 1 results)" in result
-        assert "Application library dependencies (1 of 1 results)" in result
-
-    def test_no_results_returns_message(self):
-        """When no files match, the no-results message is returned."""
-        searcher = _build_mock_searcher([
-            ("src/main/java/App.java", "no imports here"),
-        ])
-        patterns = _make_import_pattern("nonexistent.package")
-
-        result = analyze_imports(searcher, patterns, "nonexistent.package",
-                                 max_files=10, ecosystem_label="java")
-
-        assert "No imports of 'nonexistent.package' found" in result
-        assert "java" in result
-
-    def test_max_files_zero(self):
-        """max_files=0 returns headers but no individual results."""
-        package = "com.example.lib"
-        entries = [
-            ("src/main/java/App.java", f"import {package};\nuse();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=0)
-
-        assert "Main application (0 of 1 results)" in result
-        assert "Application library dependencies (0 of 0 results)" in result
-
-    def test_only_dep_results(self):
-        """When there are only dep results and no app results, full budget
-        goes to dep."""
-        package = "com.example.lib"
-        entries = [
-            ("dependencies-sources/lib-1.0-sources/Dep1.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep2.java", f"import {package};\nuse();"),
-            ("dependencies-sources/lib-1.0-sources/Dep3.java", f"import {package};\nuse();"),
-        ]
-        searcher = _build_mock_searcher(entries)
-        patterns = _make_import_pattern(package)
-
-        result = analyze_imports(searcher, patterns, package, max_files=2)
-
-        assert "Main application (0 of 0 results)" in result
-        assert "Application library dependencies (2 of 3 results)" in result
-
-
-# ---------------------------------------------------------------------------
-# A-H19: Verify tantivy query parameter is forwarded correctly
-# ---------------------------------------------------------------------------
-
-@patch("vuln_analysis.tools.import_usage_analyzer.tantivy")
-def test_analyze_imports_passes_tantivy_all_query_to_searcher(mock_tantivy):
-    """The searcher.search() call must receive the object returned by
-    tantivy.Query.all_query(), not an arbitrary value."""
-    sentinel = object()
-    mock_tantivy.Query.all_query.return_value = sentinel
-
-    docs = [_make_doc("src/App.java", "import com.example.lib;\nlib.use();")]
-    searcher = _make_searcher(docs)
-    patterns = [re.compile(re.escape("com.example.lib"), re.IGNORECASE)]
-
-    analyze_imports(searcher, patterns, "com.example.lib")
-
-    searcher.search.assert_called_once_with(sentinel, limit=1)
-
-
-# ---------------------------------------------------------------------------
-# B-M67: _find_usage_in_file doesn't skip comment lines
-# ---------------------------------------------------------------------------
-
-class TestFindUsageInFileCommentLines:
-    """Expose that _find_usage_in_file counts comment-line usages."""
-
-    def test_comment_lines_are_counted_as_usages(self):
-        """Lines starting with comment markers (// or #) that contain the
-        imported name are currently reported as usages because
-        _find_usage_in_file only skips import/from/#include prefixes."""
-        from vuln_analysis.tools.import_usage_analyzer import _find_usage_in_file
-
-        content = (
-            "import mylib\n"
-            "// mylib.init() is called on startup\n"
-            "# mylib fallback\n"
-            "mylib.doWork()\n"
-        )
-        usages = _find_usage_in_file(content, ["mylib"])
-
-        # Comment lines ARE included (this documents current behaviour)
-        usage_text = "\n".join(usages)
-        assert "mylib.doWork()" in usage_text
-        assert "// mylib.init()" in usage_text
-        assert "# mylib fallback" in usage_text
