--- src/vuln_analysis/tools/tests/test_source_code_git_loader.py
@@ -1,76 +1,0 @@
-import pytest
-import shutil
-
-from exploit_iq_commons.utils.source_code_git_loader import SourceCodeGitLoader
-
-
-@pytest.mark.parametrize("https_url,expected_ssh", [
-    (
-        "https://github.com/org/repo.git",
-        "git@github.com:org/repo.git",
-    ),
-    (
-        "https://github.com/org/repo",
-        "git@github.com:org/repo",
-    ),
-    (
-        "https://gitlab.com/group/subgroup/repo.git",
-        "git@gitlab.com:group/subgroup/repo.git",
-    ),
-    (
-        "git@github.com:org/repo.git",
-        "git@github.com:org/repo.git",  # already SSH — unchanged
-    ),
-])
-def test_https_to_ssh_url(https_url, expected_ssh):
-    assert SourceCodeGitLoader._https_to_ssh_url(https_url) == expected_ssh
-
-
-
-def test_public_clone_fails_instead_of_prompting(tmp_path):
-    """Unauthenticated clone of a missing/private repo must fail fast, not hang on a prompt."""
-    clone_url = "https://github.com/example/missing"
-    repo_path = tmp_path / "repo"
-
-    loader = SourceCodeGitLoader(repo_path=repo_path, clone_url=clone_url, ref="main")
-
-    with pytest.raises(ValueError, match=r"Failed to clone repository .+missing"):
-        loader.load_repo()
-
-    assert not repo_path.exists() or not (repo_path / ".git").exists()
-
-
-
-def test_load_repo_branch_checkout(tmp_path):
-    """Branch names must be checkable after a shallow clone (APPENG-4896)."""
-    repo_url = "https://github.com/medik8s/node-remediation-console"
-    repo_path = tmp_path / "branch_checkout_test"
-    branch = "release-0.6"
-
-    loader = SourceCodeGitLoader(repo_path=repo_path, clone_url=repo_url, ref=branch)
-    repo = loader.load_repo()
-    assert repo.head.commit.hexsha == repo.commit(branch).hexsha
-
-    shutil.rmtree(repo_path)
-
-
-
-@pytest.mark.parametrize("refs", [
-    ("3.1.43", "3.1.42", "tag switch"),
-    ("ba5c10d3655c4fec714294cbc2ae0829c44dc046",
-     "5b098ff41266d6990237cbade5b6e27cab23557f", "sha switch"),
-])
-def test_load_repo_ref_switch(tmp_path, refs):
-    repo_url = "https://github.com/gitpython-developers/GitPython"
-    repo_path = tmp_path / "git_python_test_repo"
-    ref_a, ref_b, desc = refs
-
-    loader_a = SourceCodeGitLoader(repo_path=repo_path, clone_url=repo_url, ref=ref_a)
-    repo_a = loader_a.load_repo()
-    assert repo_a.head.commit.hexsha == repo_a.commit(ref_a).hexsha, f"Initial {desc} failed"
-
-    loader_b = SourceCodeGitLoader(repo_path=repo_path, clone_url=repo_url, ref=ref_b)
-    repo_b = loader_b.load_repo()
-    assert repo_b.head.commit.hexsha == repo_b.commit(ref_b).hexsha, f"Ref switch {desc} failed"
-
-    shutil.rmtree(repo_path)
