--- src/vuln_analysis/utils/tests/test_transitive_detection.py
@@ -1,59 +1,0 @@
-import pytest
-from pathlib import Path
-
-from exploit_iq_commons.utils.dep_tree import detect_ecosystem, Ecosystem
-
-
-class TestDetectEcosystem:
-
-    def _make_repo(self, tmp_path, files):
-        for name in files:
-            p = tmp_path / name
-            p.parent.mkdir(parents=True, exist_ok=True)
-            p.write_text(files[name] if isinstance(files, dict) else "")
-        return tmp_path
-
-    def test_setup_cfg_detects_python(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"setup.cfg": "[options]\npython_requires = >=3.8\n"})
-        assert detect_ecosystem(repo) == Ecosystem.PYTHON
-
-    def test_uv_lock_detects_python(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"uv.lock": "version = 1\n"})
-        assert detect_ecosystem(repo) == Ecosystem.PYTHON
-
-    def test_poetry_lock_detects_python(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"poetry.lock": "[[package]]\n"})
-        assert detect_ecosystem(repo) == Ecosystem.PYTHON
-
-    def test_pipfile_detects_python(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"Pipfile": "[packages]\nrequests = \"*\"\n"})
-        assert detect_ecosystem(repo) == Ecosystem.PYTHON
-
-    def test_configure_with_c_file_detects_c_cpp(self, tmp_path):
-        repo = self._make_repo(tmp_path, {
-            "configure": "#!/bin/sh\n",
-            "src/main.c": "int main() { return 0; }\n",
-        })
-        assert detect_ecosystem(repo) == Ecosystem.C_CPP
-
-    def test_configure_without_c_file_returns_none(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"configure": "#!/bin/sh\n"})
-        assert detect_ecosystem(repo) is None
-
-    def test_go_mod_takes_priority_over_package_json(self, tmp_path):
-        repo = self._make_repo(tmp_path, {
-            "go.mod": "module example.com/foo\n",
-            "package.json": '{"name": "foo"}\n',
-        })
-        assert detect_ecosystem(repo) == Ecosystem.GO
-
-    def test_javascript_vs_java_priority_js_wins(self, tmp_path):
-        repo = self._make_repo(tmp_path, {
-            "package.json": '{"name": "foo"}\n',
-            "pom.xml": "<project></project>\n",
-        })
-        assert detect_ecosystem(repo) == Ecosystem.JAVASCRIPT
-
-    def test_no_manifest_returns_none(self, tmp_path):
-        repo = self._make_repo(tmp_path, {"README.md": "hello\n"})
-        assert detect_ecosystem(repo) is None
