--- src/vuln_analysis/utils/tests/test_version_check.py
@@ -1,330 +1,0 @@
-import pytest
-from unittest.mock import MagicMock
-
-from exploit_iq_commons.data_models.cve_intel import CveIntel, CveIntelGhsa, CveIntelNvd, CveIntelRhsa
-from vuln_analysis.utils.version_check import (
-    classify_version_check,
-    deterministic_version_check,
-    get_cve_description,
-    HardPathReason,
-    VersionCheckPath,
-)
-
-
-class TestClassifyEmptyOrUnknownInstalledVersion:
-
-    def test_empty_installed_version_returns_hard(self):
-        result = classify_version_check(
-            installed_version="",
-            version_info={"first_patched": "1.2.3"},
-            description=None,
-            ecosystem="maven",
-            package_name="some:pkg",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.NO_STRUCTURED_BOUNDS
-
-    def test_unknown_installed_version_returns_hard(self):
-        result = classify_version_check(
-            installed_version="unknown",
-            version_info={"first_patched": "1.2.3"},
-            description=None,
-            ecosystem="maven",
-            package_name="some:pkg",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.NO_STRUCTURED_BOUNDS
-
-
-class TestClassifyAmbiguousIntel:
-
-    def test_vulnerable_range_present_returns_ambiguous(self):
-        result = classify_version_check(
-            installed_version="1.0.0",
-            version_info={"vulnerable_range": ">= 1.0, < 2.0"},
-            description=None,
-            ecosystem="pypi",
-            package_name="requests",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.AMBIGUOUS_INTEL
-
-    def test_first_patched_and_nvd_range_both_present_returns_ambiguous(self):
-        result = classify_version_check(
-            installed_version="1.0.0",
-            version_info={
-                "first_patched": "2.0.0",
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            description=None,
-            ecosystem="pypi",
-            package_name="requests",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.AMBIGUOUS_INTEL
-
-
-class TestDeterministicNvdRangeCheck:
-
-    def test_version_inside_range_is_vulnerable(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="1.5.0",
-            version_info={
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is True
-        assert "in NVD range" in reason
-
-    def test_version_at_start_inclusive_is_vulnerable(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="1.0.0",
-            version_info={
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is True
-
-    def test_version_at_end_exclusive_is_not_vulnerable(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="2.0.0",
-            version_info={
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is False
-        assert "outside" in reason
-
-    def test_version_above_range_is_not_vulnerable(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="3.0.0",
-            version_info={
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is False
-
-    def test_version_below_range_is_not_vulnerable(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="0.5.0",
-            version_info={
-                "version_start_incl": "1.0.0",
-                "version_end_excl": "2.0.0",
-            },
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is False
-
-
-class TestGetCveDescription:
-
-    def test_nvd_description_used_first(self):
-        intel = CveIntel(
-            vuln_id="CVE-2024-0001",
-            nvd=CveIntelNvd(cve_id="CVE-2024-0001", cve_description="NVD desc"),
-            ghsa=CveIntelGhsa(ghsa_id="GHSA-xxxx", summary="GHSA summary", description="GHSA desc"),
-        )
-        result = get_cve_description(intel)
-        assert "NVD desc" in result
-        assert "GHSA desc" not in result
-
-    def test_ghsa_description_fallback_when_no_nvd(self):
-        intel = CveIntel(
-            vuln_id="CVE-2024-0002",
-            ghsa=CveIntelGhsa(ghsa_id="GHSA-xxxx", description="GHSA desc"),
-        )
-        result = get_cve_description(intel)
-        assert result == "GHSA desc"
-
-    def test_ghsa_summary_fallback_when_no_nvd_and_no_ghsa_description(self):
-        intel = CveIntel(
-            vuln_id="CVE-2024-0003",
-            ghsa=CveIntelGhsa(ghsa_id="GHSA-xxxx", summary="GHSA summary only"),
-        )
-        result = get_cve_description(intel)
-        assert result == "GHSA summary only"
-
-    def test_rhsa_details_fallback_when_no_nvd_or_ghsa(self):
-        intel = CveIntel(
-            vuln_id="CVE-2024-0004",
-            rhsa=CveIntelRhsa(
-                bugzilla=CveIntelRhsa.Bugzilla(),
-                details=["Detail 1", "Detail 2"],
-            ),
-        )
-        result = get_cve_description(intel)
-        assert "Detail 1 Detail 2" in result
-
-    def test_rhsa_statement_appended_to_nvd_description(self):
-        intel = CveIntel(
-            vuln_id="CVE-2024-0005",
-            nvd=CveIntelNvd(cve_id="CVE-2024-0005", cve_description="NVD desc"),
-            rhsa=CveIntelRhsa(
-                bugzilla=CveIntelRhsa.Bugzilla(),
-                statement="Only affects RC versions",
-            ),
-        )
-        result = get_cve_description(intel)
-        assert "NVD desc" in result
-        assert "Only affects RC versions" in result
-
-    def test_no_intel_sources_returns_empty_string(self):
-        intel = CveIntel(vuln_id="CVE-2024-0006")
-        result = get_cve_description(intel)
-        assert result == ""
-
-
-class TestClassifyEasyPath:
-    """Verify classify_version_check returns EASY for well-formed inputs."""
-
-    def test_first_patched_pypi_returns_easy(self):
-        result = classify_version_check(
-            installed_version="1.0.0",
-            version_info={"first_patched": "1.2.0"},
-            description=None,
-            ecosystem="pypi",
-            package_name="requests",
-        )
-        assert result.path == VersionCheckPath.EASY
-        assert result.hard_reason is None
-
-    def test_nvd_range_npm_returns_easy(self):
-        result = classify_version_check(
-            installed_version="4.17.20",
-            version_info={
-                "version_start_incl": "4.0.0",
-                "version_end_excl": "4.17.21",
-            },
-            description=None,
-            ecosystem="npm",
-            package_name="lodash",
-        )
-        assert result.path == VersionCheckPath.EASY
-        assert result.hard_reason is None
-
-    def test_first_patched_maven_returns_easy(self):
-        result = classify_version_check(
-            installed_version="1.9.4",
-            version_info={"first_patched": "1.9.5"},
-            description=None,
-            ecosystem="maven",
-            package_name="commons-beanutils:commons-beanutils",
-        )
-        assert result.path == VersionCheckPath.EASY
-        assert result.hard_reason is None
-
-    def test_first_patched_go_returns_easy(self):
-        result = classify_version_check(
-            installed_version="v1.21.0",
-            version_info={"first_patched": "v1.21.5"},
-            description=None,
-            ecosystem="go",
-            package_name="golang.org/x/crypto",
-        )
-        assert result.path == VersionCheckPath.EASY
-        assert result.hard_reason is None
-
-    def test_same_dist_rpm_first_patched_returns_easy(self):
-        """RPM with matching dist tag and parsable versions uses EASY path."""
-        result = classify_version_check(
-            installed_version="7.76.1-26.el9",
-            version_info={"first_patched": "7.76.1-31.el9"},
-            description=None,
-            ecosystem="rpm",
-            package_name="curl",
-        )
-        assert result.path == VersionCheckPath.EASY
-        assert result.hard_reason is None
-
-
-class TestDeterministicFirstPatchedCheck:
-    """Verify deterministic_version_check for first_patched comparison mode."""
-
-    def test_vulnerable_below_first_patched(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="1.0.0",
-            version_info={"first_patched": "1.2.0"},
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is True
-        assert "< first_patched" in reason
-
-    def test_not_vulnerable_at_first_patched(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="1.2.0",
-            version_info={"first_patched": "1.2.0"},
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is False
-        assert ">= first_patched" in reason
-
-    def test_not_vulnerable_above_first_patched(self):
-        is_vulnerable, reason = deterministic_version_check(
-            installed_version="2.0.0",
-            version_info={"first_patched": "1.2.0"},
-            ecosystem="pypi",
-        )
-        assert is_vulnerable is False
-        assert ">= first_patched" in reason
-
-
-class TestClassifyRpmComparatorDisagreement:
-    """RPM versions where GenericVersion and RpmVersion comparators disagree
-    should be routed to the HARD path with COMPARATOR_DISAGREEMENT reason."""
-
-    def test_rpm_comparator_disagreement_returns_hard(self):
-        """Epoch-style RPM versions can cause GenericVersion and RpmVersion
-        to order differently, triggering the COMPARATOR_DISAGREEMENT path."""
-        # RpmVersion handles release segments differently from GenericVersion.
-        # Find a pair where ordering disagrees: e.g., "1.0-2.el9" vs "1.0-10.el9"
-        # GenericVersion may compare "2" vs "10" lexicographically ("2" > "10"),
-        # while RpmVersion compares them numerically (2 < 10).
-        result = classify_version_check(
-            installed_version="1.0-2.el9",
-            version_info={"first_patched": "1.0-10.el9"},
-            description=None,
-            ecosystem="rpm",
-            package_name="somepkg",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.COMPARATOR_DISAGREEMENT
-
-
-class TestClassifyNoStructuredBounds:
-    """Verify HARD path when no version bounds are available."""
-
-    def test_no_version_info_returns_hard(self):
-        result = classify_version_check(
-            installed_version="1.0.0",
-            version_info={},
-            description=None,
-            ecosystem="pypi",
-            package_name="requests",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.NO_STRUCTURED_BOUNDS
-
-    def test_all_na_version_info_returns_hard(self):
-        result = classify_version_check(
-            installed_version="1.0.0",
-            version_info={
-                "first_patched": "N/A",
-                "vulnerable_range": "N/A",
-                "version_start_incl": "N/A",
-                "version_end_excl": "N/A",
-            },
-            description=None,
-            ecosystem="pypi",
-            package_name="requests",
-        )
-        assert result.path == VersionCheckPath.HARD
-        assert result.hard_reason == HardPathReason.NO_STRUCTURED_BOUNDS
