+++ tests/test_configuration_scanner.py
@@ -21,13 +21,86 @@ import pytest
 import pytest_asyncio
 
 from vuln_analysis.tools.configuration_scanner import (
+    _BINARY_EXTENSIONS,
+    _CONFIG_DIR_ALLOWED_EXTENSIONS,
     _is_config_file,
     _is_in_config_dir,
     _collect_config_files,
+    _count_config_matches,
+    format_context_snippet,
     search_config_content,
     ConfigurationScannerToolConfig,
     configuration_scanner,
 )
+
+
+class TestFormatContextSnippet:
+    def test_match_at_first_line(self):
+        lines = ["alpha", "bravo", "charlie", "delta", "echo"]
+        result = format_context_snippet(lines, match_line=0, context_lines=2)
+        output_lines = result.split("\n")
+        assert output_lines[0] == "> 1: alpha"
+        assert output_lines[1] == "  2: bravo"
+        assert output_lines[2] == "  3: charlie"
+        assert len(output_lines) == 3
+
+    def test_match_at_last_line(self):
+        lines = ["alpha", "bravo", "charlie", "delta", "echo"]
+        result = format_context_snippet(lines, match_line=4, context_lines=2)
+        output_lines = result.split("\n")
+        assert output_lines[0] == "  3: charlie"
+        assert output_lines[1] == "  4: delta"
+        assert output_lines[2] == "> 5: echo"
+        assert len(output_lines) == 3
+
+    def test_zero_context_lines(self):
+        lines = ["alpha", "bravo", "charlie"]
+        result = format_context_snippet(lines, match_line=1, context_lines=0)
+        assert result == "> 2: bravo"
+
+    def test_middle_match(self):
+        lines = ["one", "two", "three", "four", "five"]
+        result = format_context_snippet(lines, match_line=2, context_lines=1)
+        output_lines = result.split("\n")
+        assert output_lines[0] == "  2: two"
+        assert output_lines[1] == "> 3: three"
+        assert output_lines[2] == "  4: four"
+
+    def test_single_line_file(self):
+        assert format_context_snippet(["only"], match_line=0, context_lines=5) == "> 1: only"
+
+
+class TestCountConfigMatches:
+    def test_counts_matching_lines(self):
+        config_files = [
+            ("app.properties", "server.port=8080\nssl.enabled=true\nlogging.level=INFO"),
+            ("db.yml", "host: localhost\nport: 5432\nssl: false"),
+        ]
+        assert _count_config_matches(config_files, ["ssl"]) == 2
+
+    def test_case_insensitive(self):
+        config_files = [("app.yml", "SSL_ENABLED=true\nSsl_Mode=require")]
+        assert _count_config_matches(config_files, ["ssl"]) == 2
+
+    def test_no_matches(self):
+        config_files = [("app.yml", "host: localhost\nport: 8080")]
+        assert _count_config_matches(config_files, ["ssl", "tls"]) == 0
+
+    def test_multiple_keywords_same_line(self):
+        config_files = [("app.conf", "ssl_port=443")]
+        assert _count_config_matches(config_files, ["ssl", "port"]) == 1
+
+    def test_empty_config_files(self):
+        assert _count_config_matches([], ["ssl"]) == 0
+
+
+class TestBinaryExtensions:
+    def test_contains_expected_extensions(self):
+        expected = {".jar", ".png", ".jpg", ".pdf", ".class", ".zip", ".pyc"}
+        assert expected.issubset(_BINARY_EXTENSIONS)
+
+    def test_is_frozenset(self):
+        assert isinstance(_BINARY_EXTENSIONS, frozenset)
 
 
 class TestIsConfigFile:
@@ -66,8 +139,8 @@ class TestIsConfigFile:
             ("package.json", False),
             ("requirements.txt", False),
             ("setup.py", False),
-            ("setup.cfg", False),
-            ("pyproject.toml", False),
+            ("setup.cfg", True),
+            ("pyproject.toml", True),
             # Removed: catch-all extensions (only matched inside config dirs)
             ("app.yml", False),
             ("app.cfg", False),
@@ -142,6 +215,103 @@ class TestIsInConfigDir:
         assert _is_in_config_dir("application.yml") is False
 
 
+class TestConfigDirAllowedExtensions:
+    """Files in config directories are filtered by an extension allowlist."""
+
+    def test_js_in_resources_dir_skipped(self, tmp_path):
+        """Reproduces Keycloak issue: .js/.map/.css files under resources/ were collected."""
+        res_dir = tmp_path / "src" / "main" / "resources" / "theme"
+        res_dir.mkdir(parents=True)
+        for fname in ["react-core.js", "react-core.js.map", "app.css"]:
+            (res_dir / fname).write_text("not a config")
+        (res_dir / "settings.yml").write_text("key: value")
+
+        results = _collect_config_files(str(tmp_path))
+        names = [p.split("/")[-1] for p, _ in results]
+        assert "settings.yml" in names
+        assert "react-core.js" not in names
+        assert "react-core.js.map" not in names
+        assert "app.css" not in names
+
+    def test_java_source_in_resources_dir_skipped(self, tmp_path):
+        res_dir = tmp_path / "resources"
+        res_dir.mkdir()
+        (res_dir / "Helper.java").write_text("public class Helper {}")
+        names = [p.split("/")[-1] for p, _ in _collect_config_files(str(tmp_path))]
+        assert "Helper.java" not in names
+
+    def test_named_config_bypasses_extension_filter(self, tmp_path):
+        """application.properties matched by name pattern bypasses the extension filter."""
+        res_dir = tmp_path / "resources"
+        res_dir.mkdir()
+        (res_dir / "application.properties").write_text("server.port=8080")
+        names = [p.split("/")[-1] for p, _ in _collect_config_files(str(tmp_path))]
+        assert "application.properties" in names
+
+    def test_extensionless_file_in_config_dir_collected(self, tmp_path):
+        conf_dir = tmp_path / "config"
+        conf_dir.mkdir()
+        (conf_dir / "Caddyfile").write_text(":8080 { respond 200 }")
+        names = [p.split("/")[-1] for p, _ in _collect_config_files(str(tmp_path))]
+        assert "Caddyfile" in names
+
+    def test_all_allowed_extensions_collected(self, tmp_path):
+        conf_dir = tmp_path / "conf"
+        conf_dir.mkdir()
+        for ext in _CONFIG_DIR_ALLOWED_EXTENSIONS:
+            (conf_dir / f"test{ext}").write_text("content")
+        names = {p.split("/")[-1] for p, _ in _collect_config_files(str(tmp_path))}
+        for ext in _CONFIG_DIR_ALLOWED_EXTENSIONS:
+            assert f"test{ext}" in names, f"Expected test{ext} to be collected"
+
+    def test_allowlist_is_frozenset(self):
+        assert isinstance(_CONFIG_DIR_ALLOWED_EXTENSIONS, frozenset)
+
+    def test_contains_core_extensions(self):
+        expected = {".xml", ".yml", ".yaml", ".json", ".toml", ".properties", ".conf", ".ini", ".env"}
+        assert expected.issubset(_CONFIG_DIR_ALLOWED_EXTENSIONS)
+
+    def test_skips_binary_files_in_config_dir(self, tmp_path):
+        conf_dir = tmp_path / "config"
+        conf_dir.mkdir()
+        (conf_dir / "settings.yml").write_text("key: value")
+        (conf_dir / "logo.png").write_bytes(b"\x89PNG fake")
+        (conf_dir / "lib.jar").write_bytes(b"PK fake jar")
+        names = [p.split("/")[-1] for p, _ in _collect_config_files(str(tmp_path))]
+        assert "settings.yml" in names
+        assert "logo.png" not in names
+        assert "lib.jar" not in names
+
+
+class TestConfigFilePatterns:
+    """Tests for CONFIG_FILE_PATTERNS name-based matching."""
+
+    def test_build_tool_config_files(self):
+        assert _is_config_file("pyproject.toml") is True
+        assert _is_config_file("setup.cfg") is True
+        assert _is_config_file("tox.ini") is True
+        assert _is_config_file("tsconfig.json") is True
+        assert _is_config_file(".eslintrc.json") is True
+        assert _is_config_file("Makefile") is True
+        assert _is_config_file("CMakeLists.txt") is True
+        assert _is_config_file("meson.build") is True
+
+    def test_dependency_manifests_not_matched(self):
+        assert _is_config_file("go.mod") is False
+        assert _is_config_file("go.sum") is False
+        assert _is_config_file("requirements.txt") is False
+        assert _is_config_file("package.json") is False
+        assert _is_config_file("Cargo.toml") is False
+        assert _is_config_file("Gemfile") is False
+        assert _is_config_file("pom.xml") is False
+
+    def test_source_code_not_matched(self):
+        assert _is_config_file("main.java") is False
+        assert _is_config_file("app.js") is False
+        assert _is_config_file("util.py") is False
+        assert _is_config_file("handler.go") is False
+
+
 class TestCollectConfigFiles:
     def test_finds_config_files(self, tmp_path):
         (tmp_path / "config.yaml").write_text("key: value")
@@ -162,15 +332,15 @@ class TestCollectConfigFiles:
     def test_finds_files_in_config_dir(self, tmp_path):
         config_dir = tmp_path / "config"
         config_dir.mkdir()
-        (config_dir / "database.txt").write_text("db_config")
-        (config_dir / "server.txt").write_text("server_config")
+        (config_dir / "database.xml").write_text("<db/>")
+        (config_dir / "server.yml").write_text("port: 8080")
 
         result = _collect_config_files(str(tmp_path))
 
         assert len(result) == 2
         paths = {path for path, _ in result}
-        assert "config/database.txt" in paths
-        assert "config/server.txt" in paths
+        assert "config/database.xml" in paths
+        assert "config/server.yml" in paths
 
     def test_excludes_git_dir(self, tmp_path):
         git_dir = tmp_path / ".git"
@@ -259,7 +429,7 @@ class TestCollectConfigFiles:
 
         conf_dir = tmp_path / "conf"
         conf_dir.mkdir()
-        (conf_dir / "server.txt").write_text("server config")
+        (conf_dir / "server.yml").write_text("port: 8080")
 
         (tmp_path / "config.yaml").write_text("key: value")
 
@@ -268,7 +438,7 @@ class TestCollectConfigFiles:
         assert len(result) == 3
         paths = {path for path, _ in result}
         assert "src/main/resources/application.yml" in paths
-        assert "conf/server.txt" in paths
+        assert "conf/server.yml" in paths
         assert "config.yaml" in paths
 
     def test_handles_read_errors_gracefully(self, tmp_path):
@@ -569,7 +739,7 @@ class TestArun:
         # Dependency-level config
         dep_dir = tmp_path / "dependencies-sources" / "xstream-1.4" / "config"
         dep_dir.mkdir(parents=True)
-        (dep_dir / "settings.txt").write_text("xstream: dep-level")
+        (dep_dir / "settings.yml").write_text("xstream: dep-level")
 
         self._set_context(tmp_path)
