+++ tests/test_transitive_detection.py
@@ -120,3 +120,35 @@ def test_configure_in_auto_subdir(tmp_path):
     (tmp_path / "src").mkdir()
     (tmp_path / "src" / "main.c").touch()
     assert detect_ecosystem(tmp_path) == Ecosystem.C_CPP
+
+
+# --- Additional manifest detection ---
+
+def test_uv_lock_detects_python(tmp_path):
+    (tmp_path / "uv.lock").write_text("version = 1\n")
+    assert detect_ecosystem(tmp_path) == Ecosystem.PYTHON
+
+
+def test_poetry_lock_detects_python(tmp_path):
+    (tmp_path / "poetry.lock").write_text("[[package]]\n")
+    assert detect_ecosystem(tmp_path) == Ecosystem.PYTHON
+
+
+def test_configure_with_c_file_detects_c_cpp(tmp_path):
+    """Top-level 'configure' script with C source files detects C/C++."""
+    (tmp_path / "configure").write_text("#!/bin/sh\n")
+    (tmp_path / "src").mkdir()
+    (tmp_path / "src" / "main.c").write_text("int main() { return 0; }\n")
+    assert detect_ecosystem(tmp_path) == Ecosystem.C_CPP
+
+
+def test_configure_without_c_file_returns_none(tmp_path):
+    """Top-level 'configure' script without C source files is not enough."""
+    (tmp_path / "configure").write_text("#!/bin/sh\n")
+    assert detect_ecosystem(tmp_path) is None
+
+
+def test_go_mod_takes_priority_over_package_json(tmp_path):
+    (tmp_path / "go.mod").write_text("module example.com/foo\n")
+    (tmp_path / "package.json").write_text('{"name": "foo"}\n')
+    assert detect_ecosystem(tmp_path) == Ecosystem.GO
