+++ tests/test_vulnerability_intel_sanitizer.py
@@ -2,12 +2,15 @@
 # SPDX-License-Identifier: Apache-2.0
 
 """Tests for VulnerabilityIntelSanitizer — case-insensitive matching, chained
-application, and boundary conditions."""
+application, individual filter rules, and boundary conditions."""
+
+from unittest.mock import MagicMock
 
 from exploit_iq_commons.data_models.checker_status import VulnerabilityIntel
 from vuln_analysis.functions.code_agent_graph_defs import ParsedPatch, PatchFile
 from vuln_analysis.utils.vulnerability_intel_sanitizer import (
     VulnerabilityIntelSanitizer,
+    _patch_basenames,
 )
 
 
@@ -111,3 +114,175 @@ class TestEmptyIntel:
         assert result.vulnerable_patterns == []
         assert result.fix_patterns == []
         assert result.root_cause == ""
+
+
+def _patch_with_util_c() -> ParsedPatch:
+    return ParsedPatch(
+        patch_filename="cve.patch",
+        files=[
+            PatchFile(
+                source_path="tar/util.c",
+                target_path="tar/util.c",
+                hunks=[],
+            )
+        ],
+    )
+
+
+class TestSanitizeAffectedFiles:
+    def test_clears_affected_files_when_no_patch(self):
+        raw = VulnerabilityIntel(affected_files=["generator.c", "tar/util.c"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.affected_files == []
+
+    def test_clears_affected_files_when_empty_files(self):
+        patch = ParsedPatch(patch_filename="empty.patch", files=[])
+        raw = VulnerabilityIntel(affected_files=["util.c"])
+        result = VulnerabilityIntelSanitizer(patch).apply(raw)
+        assert result.affected_files == []
+
+    def test_keeps_matching_basename_with_patch(self):
+        raw = VulnerabilityIntel(affected_files=["util.c"])
+        result = VulnerabilityIntelSanitizer(_patch_with_util_c()).apply(raw)
+        assert result.affected_files == ["util.c"]
+
+    def test_keeps_full_path_when_basename_matches_patch(self):
+        raw = VulnerabilityIntel(affected_files=["tar/util.c"])
+        result = VulnerabilityIntelSanitizer(_patch_with_util_c()).apply(raw)
+        assert result.affected_files == ["tar/util.c"]
+
+    def test_drops_non_patch_file_with_patch(self):
+        raw = VulnerabilityIntel(affected_files=["util.c", "generator.c"])
+        result = VulnerabilityIntelSanitizer(_patch_with_util_c()).apply(raw)
+        assert result.affected_files == ["util.c"]
+
+
+class TestFilterVulnerableFunctions:
+    def test_drops_function_with_spaces(self):
+        raw = VulnerabilityIntel(
+            vulnerable_functions=["parse_header", "rsync compares file checksums"],
+        )
+        result = VulnerabilityIntelSanitizer(_patch_with_util_c()).apply(raw)
+        assert result.vulnerable_functions == ["parse_header"]
+
+    def test_drops_function_with_spaces_regardless_of_patch(self):
+        raw = VulnerabilityIntel(vulnerable_functions=["rsync compares file checksums"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == []
+
+    def test_keeps_function_with_tab(self):
+        raw = VulnerabilityIntel(vulnerable_functions=["parse\theader"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == ["parse\theader"]
+
+    def test_keeps_function_with_newline(self):
+        raw = VulnerabilityIntel(vulnerable_functions=["parse\nheader"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == ["parse\nheader"]
+
+    def test_keeps_function_without_whitespace(self):
+        raw = VulnerabilityIntel(vulnerable_functions=["parseHeader"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == ["parseHeader"]
+
+    def test_drops_function_with_multiple_spaces(self):
+        raw = VulnerabilityIntel(
+            vulnerable_functions=["rsync compares file checksums"],
+        )
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == []
+
+    def test_empty_list_returns_empty(self):
+        raw = VulnerabilityIntel(vulnerable_functions=[])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_functions == []
+
+
+class TestFilterSearchKeywords:
+    def test_drops_keyword_with_spaces_no_boolean(self):
+        raw = VulnerabilityIntel(
+            search_keywords=["s2length", "sum2", "rsync compares file checksums"],
+        )
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["s2length", "sum2"]
+
+    def test_keeps_keyword_with_or(self):
+        raw = VulnerabilityIntel(search_keywords=["s2length OR sum2"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["s2length OR sum2"]
+
+    def test_keeps_keyword_with_and(self):
+        raw = VulnerabilityIntel(search_keywords=["foo AND bar"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["foo AND bar"]
+
+    def test_keeps_keyword_with_lowercase_and(self):
+        raw = VulnerabilityIntel(search_keywords=["lock and key"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["lock and key"]
+
+    def test_keeps_keyword_with_lowercase_or(self):
+        raw = VulnerabilityIntel(search_keywords=["this or that"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["this or that"]
+
+    def test_keeps_keyword_with_mixed_case(self):
+        raw = VulnerabilityIntel(search_keywords=["foo And bar"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == ["foo And bar"]
+
+    def test_drops_keyword_with_spaces_no_operator(self):
+        raw = VulnerabilityIntel(search_keywords=["foo bar baz"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.search_keywords == []
+
+
+class TestRsyncStyleNoPatch:
+    def test_strips_hallucinated_paths_and_prose(self):
+        raw = VulnerabilityIntel(
+            affected_files=["generator.c", "src/foo.c"],
+            vulnerable_functions=["rsync compares file checksums"],
+            vulnerable_variables=["s2length", "sum2"],
+            search_keywords=["s2length", "sum2", "rsync compares file checksums"],
+        )
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.affected_files == []
+        assert result.vulnerable_functions == []
+        assert result.vulnerable_variables == ["s2length", "sum2"]
+        assert result.search_keywords == ["s2length", "sum2"]
+
+
+class TestPatchBasenames:
+    def test_empty_source_path_skipped(self):
+        patch_file = PatchFile(source_path="", target_path="tar/util.c", hunks=[])
+        parsed = ParsedPatch(patch_filename="test.patch", files=[patch_file])
+        result = _patch_basenames(parsed)
+        assert result == {"util.c"}
+
+    def test_none_target_path_skipped(self):
+        """PatchFile requires str, so use a mock to verify the None guard."""
+        mock_file = MagicMock()
+        mock_file.source_path = "src/foo.c"
+        mock_file.target_path = None
+        parsed = MagicMock()
+        parsed.files = [mock_file]
+        result = _patch_basenames(parsed)
+        assert result == {"foo.c"}
+
+    def test_normal_paths_extracted(self):
+        patch_file = PatchFile(
+            source_path="a/lib/parser.h",
+            target_path="b/lib/parser.h",
+            hunks=[],
+        )
+        parsed = ParsedPatch(patch_filename="test.patch", files=[patch_file])
+        result = _patch_basenames(parsed)
+        assert result == {"parser.h"}
+
+
+class TestVulnerableVariablesPassThrough:
+    def test_variables_with_spaces_not_filtered(self):
+        """vulnerable_variables receive no sanitization — spaces pass through."""
+        raw = VulnerabilityIntel(vulnerable_variables=["foo bar", "valid"])
+        result = VulnerabilityIntelSanitizer(None).apply(raw)
+        assert result.vulnerable_variables == ["foo bar", "valid"]
