+++ tests/test_web_patch_fetcher.py
@@ -15,14 +15,26 @@
 
 """Tests for web_patch_fetcher: _parse_patch_content, _resolve_gitiles_url, _extract_commit_from_references."""
 
-from unittest.mock import MagicMock
+import base64
 
+import aiohttp
 import pytest
+from unittest.mock import AsyncMock, MagicMock, patch
 
+from vuln_analysis.functions.code_agent_graph_defs import PatchHunk, PatchFile, ParsedPatch
+from vuln_analysis.utils.intel_utils import extract_functions_from_parsed_patch
 from vuln_analysis.utils.web_patch_fetcher import (
     _parse_patch_content,
     WebPatchFetcher,
+    WebPatchResult,
     OSVClient,
+    build_patch_url_from_repo,
+    fetch_patch_for_cve,
+    _GITHUB_COMMIT_PATTERN,
+    _GITHUB_PR_PATTERN,
+    _GITWEB_COMMIT_PATTERN,
+    _KERNEL_CGIT_COMMIT_PATTERN,
+    _KERNEL_SHORT_PATTERN,
 )
 
 
@@ -283,3 +295,1209 @@ class TestFetchGitilesPatchEncoding:
         auto_parsed = yarl.URL(patch_url)
         assert "%255E" in str(auto_parsed) or "%5E%21" not in str(auto_parsed), \
             "Without encoded=True, yarl re-encodes the percent sign"
+
+
+# ---------------------------------------------------------------------------
+# URL pattern matching tests
+# ---------------------------------------------------------------------------
+
+class TestUrlPatterns:
+    """Test URL pattern matching."""
+
+    def test_github_commit_pattern(self):
+        """Test GitHub commit URL pattern matching."""
+        url = "https://github.com/libarchive/libarchive/commit/a2a73a8f14b3208c7f6acbbc93265254a7c1efd0"
+        match = _GITHUB_COMMIT_PATTERN.match(url)
+        assert match is not None
+        assert match.group(1) == "libarchive/libarchive"
+        assert match.group(2) == "a2a73a8f14b3208c7f6acbbc93265254a7c1efd0"
+
+    def test_github_commit_pattern_short_sha(self):
+        """Test GitHub commit URL with short SHA."""
+        url = "https://github.com/curl/curl/commit/39d1976b7f"
+        match = _GITHUB_COMMIT_PATTERN.match(url)
+        assert match is not None
+        assert match.group(1) == "curl/curl"
+        assert match.group(2) == "39d1976b7f"
+
+    def test_github_pr_pattern(self):
+        """Test GitHub PR URL pattern matching."""
+        url = "https://github.com/libarchive/libarchive/pull/2934"
+        match = _GITHUB_PR_PATTERN.match(url)
+        assert match is not None
+        assert match.group(1) == "libarchive/libarchive"
+        assert match.group(2) == "2934"
+
+    def test_kernel_cgit_commit_pattern(self):
+        """Test kernel.org cgit commit URL pattern."""
+        url = (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+            "/commit/?id=f342de4e2f33e0e39165d8639387aa6c19dff660"
+        )
+        match = _KERNEL_CGIT_COMMIT_PATTERN.match(url)
+        assert match is not None
+        expected_repo = "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+        assert match.group(1) == expected_repo
+        assert match.group(2) == "f342de4e2f33e0e39165d8639387aa6c19dff660"
+
+    def test_kernel_short_pattern(self):
+        """Test kernel.org short URL pattern."""
+        url = "https://git.kernel.org/stable/c/096bb5b43edf755bc4477e64004fa3a20539ec2f"
+        match = _KERNEL_SHORT_PATTERN.match(url)
+        assert match is not None
+        assert match.group(1) == "stable"
+        assert match.group(2) == "096bb5b43edf755bc4477e64004fa3a20539ec2f"
+
+    def test_gitweb_commit_pattern(self):
+        """Test gitweb commit URL pattern."""
+        url = "https://git.samba.org/?p=rsync.git;a=commit;h=6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        match = _GITWEB_COMMIT_PATTERN.match(url)
+        assert match is not None
+        assert match.group(1) == "https://git.samba.org"
+        assert match.group(2) == "rsync.git"
+        assert match.group(3) == "6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+
+
+# ---------------------------------------------------------------------------
+# WebPatchFetcher URL resolution tests
+# ---------------------------------------------------------------------------
+
+class TestWebPatchFetcherUrlResolution:
+    """Test URL resolution in WebPatchFetcher."""
+
+    @pytest.fixture
+    def fetcher(self):
+        """Create a WebPatchFetcher with a mock session."""
+        mock_session = MagicMock()
+        return WebPatchFetcher(session=mock_session)
+
+    def test_resolve_github_commit_url(self, fetcher):
+        """Test resolving GitHub commit URL to patch URL."""
+        url = "https://github.com/curl/curl/commit/39d1976b7f"
+        resolved = fetcher._resolve_github_url(url)
+        assert resolved is not None
+        assert resolved.patch_url == "https://github.com/curl/curl/commit/39d1976b7f.patch"
+        assert resolved.platform == "github"
+        assert resolved.url_type == "commit"
+        assert resolved.repo_url == "https://github.com/curl/curl"
+        assert resolved.commit_sha == "39d1976b7f"
+
+    def test_resolve_github_pr_url(self, fetcher):
+        """Test resolving GitHub PR URL to patch URL."""
+        url = "https://github.com/libarchive/libarchive/pull/2934"
+        resolved = fetcher._resolve_github_url(url)
+        assert resolved is not None
+        assert resolved.patch_url == "https://github.com/libarchive/libarchive/pull/2934.patch"
+        assert resolved.platform == "github"
+        assert resolved.url_type == "pull"
+        assert resolved.repo_url == "https://github.com/libarchive/libarchive"
+        assert resolved.commit_sha == "PR-2934"
+
+    def test_resolve_ubuntu_style_url(self, fetcher):
+        """Test resolving Ubuntu-style 'upstream:' URL."""
+        url = "upstream: https://github.com/curl/curl/commit/39d1976b7f"
+        resolved = fetcher._resolve_github_url(url)
+        assert resolved is not None
+        assert resolved.patch_url == "https://github.com/curl/curl/commit/39d1976b7f.patch"
+        assert resolved.platform == "github"
+        assert resolved.url_type == "commit"
+        assert resolved.repo_url == "https://github.com/curl/curl"
+        assert resolved.commit_sha == "39d1976b7f"
+
+    def test_resolve_kernel_cgit_commit_url(self, fetcher):
+        """Test resolving kernel.org cgit commit URL."""
+        url = (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+            "/commit/?id=f342de4e2f33"
+        )
+        resolved = fetcher._resolve_kernel_org_url(url)
+        assert resolved is not None
+        expected_patch = (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+            "/patch/?id=f342de4e2f33"
+        )
+        assert resolved.patch_url == expected_patch
+        assert resolved.platform == "kernel.org"
+        assert resolved.url_type == "commit"
+        assert resolved.commit_sha == "f342de4e2f33"
+
+    def test_resolve_kernel_short_stable_url(self, fetcher):
+        """Test resolving kernel.org short stable URL."""
+        url = "https://git.kernel.org/stable/c/096bb5b43edf"
+        resolved = fetcher._resolve_kernel_org_url(url)
+        assert resolved is not None
+        expected_patch = (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git"
+            "/patch/?id=096bb5b43edf"
+        )
+        assert resolved.patch_url == expected_patch
+        assert resolved.platform == "kernel.org"
+        assert resolved.url_type == "commit"
+        assert resolved.commit_sha == "096bb5b43edf"
+        assert resolved.repo_url == (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git"
+        )
+
+    def test_resolve_kernel_short_torvalds_url(self, fetcher):
+        """Test resolving kernel.org short torvalds URL."""
+        url = "https://git.kernel.org/torvalds/c/abc123def456"
+        resolved = fetcher._resolve_kernel_org_url(url)
+        assert resolved is not None
+        expected_patch = (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+            "/patch/?id=abc123def456"
+        )
+        assert resolved.patch_url == expected_patch
+        assert resolved.platform == "kernel.org"
+        assert resolved.url_type == "commit"
+        assert resolved.commit_sha == "abc123def456"
+        assert resolved.repo_url == (
+            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+        )
+
+    def test_resolve_unknown_url_returns_none(self, fetcher):
+        """Test that unknown URLs return None."""
+        url = "https://example.com/some/path"
+        resolved = fetcher._resolve_to_patch_url(url)
+        assert resolved is None
+
+    def test_resolve_gitweb_commit_url(self, fetcher):
+        """Test resolving git.samba.org gitweb commit URL."""
+        url = (
+            "https://git.samba.org/?p=rsync.git;a=commit;"
+            "h=6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        )
+        resolved = fetcher._resolve_gitweb_url(url)
+        assert resolved is not None
+        assert resolved.platform == "gitweb"
+        assert resolved.patch_url == (
+            "https://git.samba.org/?p=rsync.git;a=patch;"
+            "h=6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        )
+        assert resolved.repo_url == "https://git.samba.org/rsync.git"
+        assert resolved.commit_sha == "6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+
+    def test_resolve_gitweb_patch_url(self, fetcher):
+        """Test resolving gitweb patch URL directly."""
+        url = (
+            "https://git.samba.org/?p=rsync.git;a=patch;"
+            "h=6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        )
+        resolved = fetcher._resolve_gitweb_url(url)
+        assert resolved is not None
+        assert resolved.platform == "gitweb"
+        assert resolved.patch_url == url
+
+
+# ---------------------------------------------------------------------------
+# build_patch_url_from_repo tests
+# ---------------------------------------------------------------------------
+
+class TestBuildPatchUrlFromRepo:
+    """Test patch URL construction from OSV repo + commit."""
+
+    def test_samba_rsync_gitweb(self):
+        repo = "https://git.samba.org/rsync.git"
+        sha = "6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        patch_url = build_patch_url_from_repo(repo, sha)
+        assert patch_url == f"https://git.samba.org/?p=rsync.git;a=patch;h={sha}"
+
+    def test_kernel_cgit_path_info(self):
+        repo = "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git"
+        sha = "f342de4e2f33"
+        patch_url = build_patch_url_from_repo(repo, sha)
+        assert patch_url == f"{repo}/patch/?id={sha}"
+
+    def test_github_repo(self):
+        repo = "https://github.com/openssl/openssl"
+        sha = "abc123def456"
+        patch_url = build_patch_url_from_repo(repo, sha)
+        assert patch_url == "https://github.com/openssl/openssl/commit/abc123def456.patch"
+
+
+# ---------------------------------------------------------------------------
+# Prioritization and deduplication tests
+# ---------------------------------------------------------------------------
+
+class TestPrioritizeAndDedupe:
+    """Test URL prioritization and deduplication."""
+
+    @pytest.fixture
+    def fetcher(self):
+        """Create a WebPatchFetcher with a mock session."""
+        mock_session = MagicMock()
+        return WebPatchFetcher(session=mock_session)
+
+    def test_prioritize_commit_over_pull(self, fetcher):
+        """Test that commit URLs are prioritized over PR URLs."""
+        candidates = {
+            "ghsa": [
+                "https://github.com/foo/bar/pull/123",
+                "https://github.com/foo/bar/commit/abc123",
+            ],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert len(result) == 2
+        assert result[0][2] == "commit"  # commit first
+        assert result[1][2] == "pull"  # PR second
+
+    def test_prioritize_ubuntu_patches_first(self, fetcher):
+        """Test that ubuntu_patches source is prioritized for commits."""
+        candidates = {
+            "ghsa": ["https://github.com/foo/bar/commit/ghsa123"],
+            "ubuntu_patches": ["upstream: https://github.com/foo/bar/commit/ubuntu123"],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert len(result) == 2
+        assert result[0][1] == "ubuntu_patches"  # ubuntu_patches first
+        assert result[1][1] == "ghsa"
+
+    def test_deduplicate_same_url(self, fetcher):
+        """Test that duplicate URLs are deduplicated."""
+        candidates = {
+            "ghsa": ["https://github.com/foo/bar/commit/abc123"],
+            "nvd": ["https://github.com/foo/bar/commit/abc123"],
+            "rhsa": ["https://github.com/foo/bar/commit/abc123"],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert len(result) == 1  # Only one unique URL
+
+    def test_deduplicate_with_trailing_slash(self, fetcher):
+        """Test deduplication handles trailing slashes."""
+        candidates = {
+            "ghsa": ["https://github.com/foo/bar/commit/abc123"],
+            "nvd": ["https://github.com/foo/bar/commit/abc123/"],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert len(result) == 1
+
+    def test_kernel_org_urls_prioritized_as_commits(self, fetcher):
+        """Test that kernel.org URLs are recognized as commit URLs."""
+        candidates = {
+            "nvd": [
+                "https://git.kernel.org/stable/c/abc123",
+                "https://github.com/foo/bar/pull/456",
+            ],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert len(result) == 2
+        assert result[0][2] == "commit"  # kernel.org first (commit)
+        assert result[1][2] == "pull"  # PR second
+
+    def test_empty_candidates(self, fetcher):
+        """Test handling of empty candidates."""
+        candidates = {}
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert result == []
+
+    def test_skip_non_matching_urls(self, fetcher):
+        """Test that non-commit/PR URLs are skipped."""
+        candidates = {
+            "ghsa": [
+                "https://github.com/advisories/GHSA-xxxx-xxxx-xxxx",
+                "https://example.com/something",
+            ],
+        }
+        result = fetcher._prioritize_and_dedupe(candidates)
+        assert result == []  # Advisory URLs don't match commit/PR patterns
+
+
+# ---------------------------------------------------------------------------
+# URL normalization tests
+# ---------------------------------------------------------------------------
+
+class TestNormalizeUrl:
+    """Test URL normalization for deduplication."""
+
+    @pytest.fixture
+    def fetcher(self):
+        mock_session = MagicMock()
+        return WebPatchFetcher(session=mock_session)
+
+    def test_normalize_removes_trailing_slash(self, fetcher):
+        url = "https://github.com/foo/bar/commit/abc123/"
+        normalized = fetcher._normalize_url_for_dedupe(url)
+        assert not normalized.endswith("/")
+
+    def test_normalize_lowercases(self, fetcher):
+        url = "https://GitHub.com/Foo/Bar/Commit/ABC123"
+        normalized = fetcher._normalize_url_for_dedupe(url)
+        assert normalized == normalized.lower()
+
+    def test_normalize_removes_patch_suffix(self, fetcher):
+        url = "https://github.com/foo/bar/commit/abc123.patch"
+        normalized = fetcher._normalize_url_for_dedupe(url)
+        assert not normalized.endswith(".patch")
+
+
+# ---------------------------------------------------------------------------
+# Commit URL detection tests
+# ---------------------------------------------------------------------------
+
+class TestIsCommitUrl:
+    """Test commit URL detection."""
+
+    @pytest.fixture
+    def fetcher(self):
+        mock_session = MagicMock()
+        return WebPatchFetcher(session=mock_session)
+
+    def test_github_commit_is_commit_url(self, fetcher):
+        assert fetcher._is_commit_url("https://github.com/foo/bar/commit/abc123")
+
+    def test_kernel_cgit_is_commit_url(self, fetcher):
+        assert fetcher._is_commit_url("https://git.kernel.org/pub/scm/linux.git/commit/?id=abc123")
+
+    def test_kernel_short_is_commit_url(self, fetcher):
+        assert fetcher._is_commit_url("https://git.kernel.org/stable/c/abc123")
+
+    def test_github_pr_is_not_commit_url(self, fetcher):
+        assert not fetcher._is_commit_url("https://github.com/foo/bar/pull/123")
+
+    def test_advisory_is_not_commit_url(self, fetcher):
+        assert not fetcher._is_commit_url("https://github.com/advisories/GHSA-xxxx")
+
+    def test_gitweb_commit_is_commit_url(self, fetcher):
+        url = "https://git.samba.org/?p=rsync.git;a=commit;h=6c8ca91c731b7bf2b081694bda85b7dadc2b7aff"
+        assert fetcher._is_commit_url(url)
+
+
+# ---------------------------------------------------------------------------
+# PR URL detection tests
+# ---------------------------------------------------------------------------
+
+class TestIsPrUrl:
+    """Test PR URL detection."""
+
+    @pytest.fixture
+    def fetcher(self):
+        mock_session = MagicMock()
+        return WebPatchFetcher(session=mock_session)
+
+    def test_github_pr_is_pr_url(self, fetcher):
+        assert fetcher._is_pr_url("https://github.com/foo/bar/pull/123")
+
+    def test_github_commit_is_not_pr_url(self, fetcher):
+        assert not fetcher._is_pr_url("https://github.com/foo/bar/commit/abc123")
+
+
+# ---------------------------------------------------------------------------
+# fetch_from_intel_refs integration tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestFetchFromIntelRefs:
+    """Test fetch_from_intel_refs integration."""
+
+    @pytest.fixture
+    def mock_session(self):
+        return MagicMock()
+
+    async def test_fetch_from_intel_refs_empty_candidates(self, mock_session):
+        """Test handling of empty candidates."""
+        fetcher = WebPatchFetcher(session=mock_session)
+        result = await fetcher.fetch_from_intel_refs({}, "CVE-2024-1234")
+        assert result is None
+
+    async def test_fetch_from_intel_refs_returns_first_success(self, mock_session):
+        """Test that first successful fetch is returned."""
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        # Mock the fetch_from_url method
+        with patch.object(fetcher, 'fetch_from_url') as mock_fetch:
+            mock_result = MagicMock()
+            mock_result.parsed_patch = MagicMock()
+            mock_fetch.return_value = mock_result
+
+            candidates = {
+                "ubuntu_patches": ["upstream: https://github.com/foo/bar/commit/abc123"],
+            }
+            result = await fetcher.fetch_from_intel_refs(candidates, "CVE-2024-1234")
+
+            assert result is mock_result
+            mock_fetch.assert_called_once()
+            # Verify the URL and cve_id passed to fetch_from_url
+            call_args = mock_fetch.call_args
+            assert "abc123" in call_args[0][0]  # URL contains the commit SHA
+            assert call_args[0][1] == "CVE-2024-1234"  # cve_id
+
+    async def test_fetch_from_intel_refs_tries_multiple_urls(self, mock_session):
+        """Test that multiple URLs are tried on failure."""
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch.object(fetcher, 'fetch_from_url') as mock_fetch:
+            # First call returns None (failure), second returns success
+            mock_result = MagicMock()
+            mock_result.parsed_patch = MagicMock()
+            mock_fetch.side_effect = [None, mock_result]
+
+            candidates = {
+                "ubuntu_patches": ["upstream: https://github.com/foo/bar/commit/abc123"],
+                "ghsa": ["https://github.com/foo/bar/commit/def456"],
+            }
+            result = await fetcher.fetch_from_intel_refs(candidates, "CVE-2024-1234")
+
+            assert result is mock_result
+            assert mock_fetch.call_count == 2
+            # Verify URLs were tried in priority order (ubuntu_patches first)
+            first_call_url = mock_fetch.call_args_list[0][0][0]
+            second_call_url = mock_fetch.call_args_list[1][0][0]
+            assert "abc123" in first_call_url  # ubuntu_patches commit tried first
+            assert "def456" in second_call_url  # ghsa commit tried second
+
+
+# ---------------------------------------------------------------------------
+# PatchHunk section_header tests
+# ---------------------------------------------------------------------------
+
+class TestPatchHunkSectionHeader:
+    """Tests for PatchHunk.section_header field (fix 1)."""
+
+    def test_section_header_default_empty(self):
+        hunk = PatchHunk(source_start=1, source_length=5, target_start=1, target_length=5)
+        assert hunk.section_header == ""
+
+    def test_section_header_populated(self):
+        hunk = PatchHunk(
+            source_start=10, source_length=5, target_start=10, target_length=5,
+            section_header="void processRequest(HttpServletRequest req)",
+        )
+        assert hunk.section_header == "void processRequest(HttpServletRequest req)"
+
+    def test_section_header_in_extract_functions(self):
+        """section_header containing a function signature should be picked up by extract_functions_from_parsed_patch."""
+        patch_obj = ParsedPatch(
+            patch_filename="fix.patch",
+            files=[PatchFile(
+                source_path="a/src/main.go",
+                target_path="b/src/main.go",
+                hunks=[PatchHunk(
+                    source_start=10, source_length=3, target_start=10, target_length=3,
+                    section_header="func handleRequest(w http.ResponseWriter, r *http.Request)",
+                    added_lines=["    return nil"],
+                )],
+            )],
+        )
+        funcs = extract_functions_from_parsed_patch(patch_obj, ecosystem="go")
+        assert "handleRequest" in funcs
+
+    def test_section_header_only_no_added_lines(self):
+        """Function name should be extracted from section_header even when added_lines have no definitions."""
+        patch_obj = ParsedPatch(
+            patch_filename="fix.patch",
+            files=[PatchFile(
+                source_path="a/Foo.java",
+                target_path="b/Foo.java",
+                hunks=[PatchHunk(
+                    source_start=50, source_length=2, target_start=50, target_length=2,
+                    section_header="public String getVersion()",
+                    added_lines=["        return VERSION;"],
+                )],
+            )],
+        )
+        funcs = extract_functions_from_parsed_patch(patch_obj, ecosystem="java")
+        assert "getVersion" in funcs
+
+    def test_section_header_skipped_for_test_files(self):
+        """Test files should be skipped even if section_header has a function."""
+        patch_obj = ParsedPatch(
+            patch_filename="fix.patch",
+            files=[PatchFile(
+                source_path="a/FooTest.java",
+                target_path="b/FooTest.java",
+                hunks=[PatchHunk(
+                    source_start=10, source_length=2, target_start=10, target_length=2,
+                    section_header="public void shouldParseInput()",
+                    added_lines=["        assertEquals(1, 1);"],
+                )],
+            )],
+        )
+        funcs = extract_functions_from_parsed_patch(patch_obj, ecosystem="java")
+        assert len(funcs) == 0
+
+    def test_section_header_empty_string_not_matched(self):
+        """Empty section_header should not contribute any function names."""
+        patch_obj = ParsedPatch(
+            patch_filename="fix.patch",
+            files=[PatchFile(
+                source_path="a/main.c",
+                target_path="b/main.c",
+                hunks=[PatchHunk(
+                    source_start=1, source_length=1, target_start=1, target_length=1,
+                    section_header="",
+                    added_lines=["int x = 0;"],
+                )],
+            )],
+        )
+        funcs = extract_functions_from_parsed_patch(patch_obj, ecosystem="c")
+        assert len(funcs) == 0
+
+    def test_python_section_header(self):
+        patch_obj = ParsedPatch(
+            patch_filename="fix.patch",
+            files=[PatchFile(
+                source_path="a/utils.py",
+                target_path="b/utils.py",
+                hunks=[PatchHunk(
+                    source_start=20, source_length=3, target_start=20, target_length=3,
+                    section_header="def validate_input(data):",
+                    added_lines=["    if not data:"],
+                )],
+            )],
+        )
+        funcs = extract_functions_from_parsed_patch(patch_obj, ecosystem="python")
+        assert "validate_input" in funcs
+
+
+# ---------------------------------------------------------------------------
+# PatchFile.clean_target_path tests
+# ---------------------------------------------------------------------------
+
+class TestCleanTargetPath:
+    """Tests for PatchFile.clean_target_path property (fix 3)."""
+
+    def test_strip_b_prefix(self):
+        pf = PatchFile(source_path="a/foo.c", target_path="b/foo.c", hunks=[])
+        assert pf.clean_target_path == "foo.c"
+
+    def test_strip_a_prefix(self):
+        pf = PatchFile(source_path="a/foo.c", target_path="a/foo.c", hunks=[])
+        assert pf.clean_target_path == "foo.c"
+
+    def test_no_prefix(self):
+        pf = PatchFile(source_path="foo.c", target_path="foo.c", hunks=[])
+        assert pf.clean_target_path == "foo.c"
+
+    def test_nested_path_b_prefix(self):
+        pf = PatchFile(source_path="a/src/main/Bar.java", target_path="b/src/main/Bar.java", hunks=[])
+        assert pf.clean_target_path == "src/main/Bar.java"
+
+    def test_b_in_path_not_prefix(self):
+        """A 'b/' that is part of the path (not a prefix) should not be stripped."""
+        pf = PatchFile(source_path="a/lib/b/file.py", target_path="lib/b/file.py", hunks=[])
+        assert pf.clean_target_path == "lib/b/file.py"
+
+    def test_devnull_path(self):
+        """New files have /dev/null as source — target should still strip prefix."""
+        pf = PatchFile(source_path="/dev/null", target_path="b/new_file.go", hunks=[], is_new_file=True)
+        assert pf.clean_target_path == "new_file.go"
+
+
+# ---------------------------------------------------------------------------
+# fetch_patch_for_cve orchestrator tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestFetchPatchForCve:
+    """Test the top-level fetch_patch_for_cve orchestrator."""
+
+    async def test_fetch_patch_for_cve_intel_refs_success(self):
+        """When intel refs return a result with parsed_patch, return it without hitting OSV."""
+        mock_result = MagicMock(spec=WebPatchResult)
+        mock_result.parsed_patch = MagicMock()  # non-None
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_fetcher_instance.fetch_from_intel_refs = AsyncMock(return_value=mock_result)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {"ghsa": ["https://github.com/foo/bar/commit/abc123"]},
+                "CVE-2024-1234",
+            )
+            assert result is mock_result
+            # OSVClient.get_fix_patch should never be called
+            MockOSV.return_value.get_fix_patch.assert_not_called()
+
+    async def test_fetch_patch_for_cve_osv_fallback(self):
+        """When intel refs return None, fall back to OSV."""
+        mock_osv_result = MagicMock(spec=WebPatchResult)
+        mock_osv_result.parsed_patch = MagicMock()
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_fetcher_instance.fetch_from_intel_refs = AsyncMock(return_value=None)
+            mock_osv_instance = MockOSV.return_value
+            mock_osv_instance.get_fix_patch = AsyncMock(return_value=mock_osv_result)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {"ghsa": ["https://github.com/foo/bar/commit/abc123"]},
+                "CVE-2024-1234",
+                upstream_version="1.2.3",
+                package_name="bar",
+            )
+            assert result is mock_osv_result
+            mock_osv_instance.get_fix_patch.assert_called_once_with(
+                "CVE-2024-1234", "1.2.3", "bar"
+            )
+
+    async def test_fetch_patch_for_cve_both_fail(self):
+        """When both intel refs and OSV return None, function returns None."""
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_fetcher_instance.fetch_from_intel_refs = AsyncMock(return_value=None)
+            mock_osv_instance = MockOSV.return_value
+            mock_osv_instance.get_fix_patch = AsyncMock(return_value=None)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {"ghsa": ["https://github.com/foo/bar/commit/abc123"]},
+                "CVE-2024-5555",
+            )
+            assert result is None
+
+    async def test_fetch_patch_for_cve_empty_candidates(self):
+        """Empty candidates dict skips intel refs, goes straight to OSV."""
+        mock_osv_result = MagicMock(spec=WebPatchResult)
+        mock_osv_result.parsed_patch = MagicMock()
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_osv_instance = MockOSV.return_value
+            mock_osv_instance.get_fix_patch = AsyncMock(return_value=mock_osv_result)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {},
+                "CVE-2024-9999",
+            )
+            assert result is mock_osv_result
+            # fetch_from_intel_refs should not have been called
+            mock_fetcher_instance.fetch_from_intel_refs.assert_not_called()
+
+    async def test_fetch_patch_for_cve_intel_result_without_parsed_patch(self):
+        """Intel refs returning a result with parsed_patch=None should fall back to OSV."""
+        mock_intel_result = MagicMock(spec=WebPatchResult)
+        mock_intel_result.parsed_patch = None
+
+        mock_osv_result = MagicMock(spec=WebPatchResult)
+        mock_osv_result.parsed_patch = MagicMock()
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_fetcher_instance.fetch_from_intel_refs = AsyncMock(return_value=mock_intel_result)
+            mock_osv_instance = MockOSV.return_value
+            mock_osv_instance.get_fix_patch = AsyncMock(return_value=mock_osv_result)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {"ghsa": ["https://github.com/foo/bar/commit/abc123"]},
+                "CVE-2024-7777",
+            )
+            assert result is mock_osv_result
+
+    async def test_fetch_patch_for_cve_candidates_with_empty_lists(self):
+        """Candidates dict where all lists are empty should skip intel refs."""
+        mock_osv_result = MagicMock(spec=WebPatchResult)
+        mock_osv_result.parsed_patch = MagicMock()
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.WebPatchFetcher") as MockFetcher, \
+             patch("vuln_analysis.utils.web_patch_fetcher.OSVClient") as MockOSV:
+            mock_fetcher_instance = MockFetcher.return_value
+            mock_osv_instance = MockOSV.return_value
+            mock_osv_instance.get_fix_patch = AsyncMock(return_value=mock_osv_result)
+
+            session = MagicMock(spec=aiohttp.ClientSession)
+            result = await fetch_patch_for_cve(
+                session,
+                {"ghsa": [], "nvd": [], "rhsa": []},
+                "CVE-2024-8888",
+            )
+            assert result is mock_osv_result
+            mock_fetcher_instance.fetch_from_intel_refs.assert_not_called()
+
+
+# ---------------------------------------------------------------------------
+# fetch_from_url tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestFetchFromUrl:
+    """Test WebPatchFetcher.fetch_from_url."""
+
+    async def test_fetch_from_url_success(self):
+        """Successful URL resolution and patch fetch returns WebPatchResult."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        mock_resolved = MagicMock()
+        mock_resolved.patch_url = "https://github.com/foo/bar/commit/abc123.patch"
+        mock_resolved.platform = "github"
+        mock_resolved.url_type = "commit"
+        mock_resolved.repo_url = "https://github.com/foo/bar"
+        mock_resolved.commit_sha = "abc123def456"
+
+        patch_text = (
+            "From: Dev <dev@example.com>\n"
+            "Subject: Fix bug\n\n"
+            "--- a/file.c\n"
+            "+++ b/file.c\n"
+            "@@ -1,2 +1,3 @@\n"
+            " line1\n"
+            "+line2\n"
+            " line3\n"
+        )
+
+        with patch.object(fetcher, "_resolve_to_patch_url", return_value=mock_resolved), \
+             patch.object(fetcher, "_fetch_patch_content", new_callable=AsyncMock, return_value=patch_text):
+            result = await fetcher.fetch_from_url(
+                "https://github.com/foo/bar/commit/abc123def456",
+                "CVE-2024-1234",
+                source="ghsa",
+            )
+            assert result is not None
+            assert isinstance(result, WebPatchResult)
+            assert result.cve_id == "CVE-2024-1234"
+            assert result.platform == "github"
+            assert result.source == "ghsa"
+            assert result.parsed_patch is not None
+
+    async def test_fetch_from_url_cache_hit(self):
+        """Second call with the same URL returns cached result."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        mock_resolved = MagicMock()
+        mock_resolved.patch_url = "https://github.com/foo/bar/commit/abc123.patch"
+        mock_resolved.platform = "github"
+        mock_resolved.url_type = "commit"
+        mock_resolved.repo_url = "https://github.com/foo/bar"
+        mock_resolved.commit_sha = "abc123def456"
+
+        patch_text = (
+            "--- a/file.c\n"
+            "+++ b/file.c\n"
+            "@@ -1,2 +1,3 @@\n"
+            " line1\n"
+            "+line2\n"
+            " line3\n"
+        )
+
+        with patch.object(fetcher, "_resolve_to_patch_url", return_value=mock_resolved) as mock_resolve, \
+             patch.object(fetcher, "_fetch_patch_content", new_callable=AsyncMock, return_value=patch_text) as mock_fetch:
+            # First call
+            result1 = await fetcher.fetch_from_url(
+                "https://github.com/foo/bar/commit/abc123def456",
+                "CVE-2024-1234",
+            )
+            # Second call with same URL
+            result2 = await fetcher.fetch_from_url(
+                "https://github.com/foo/bar/commit/abc123def456",
+                "CVE-2024-1234",
+            )
+            assert result1 is result2
+            # _resolve_to_patch_url should only be called once; second call hits cache
+            mock_resolve.assert_called_once()
+
+    async def test_fetch_from_url_unresolvable_url(self):
+        """When _resolve_to_patch_url returns None, fetch_from_url returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch.object(fetcher, "_resolve_to_patch_url", return_value=None):
+            result = await fetcher.fetch_from_url(
+                "https://example.com/not-a-patch-url",
+                "CVE-2024-1234",
+            )
+            assert result is None
+
+    async def test_fetch_from_url_fetch_fails(self):
+        """When _fetch_patch_content returns None, fetch_from_url returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        mock_resolved = MagicMock()
+        mock_resolved.patch_url = "https://github.com/foo/bar/commit/abc123.patch"
+        mock_resolved.platform = "github"
+        mock_resolved.url_type = "commit"
+        mock_resolved.repo_url = "https://github.com/foo/bar"
+        mock_resolved.commit_sha = "abc123def456"
+
+        with patch.object(fetcher, "_resolve_to_patch_url", return_value=mock_resolved), \
+             patch.object(fetcher, "_fetch_patch_content", new_callable=AsyncMock, return_value=None):
+            result = await fetcher.fetch_from_url(
+                "https://github.com/foo/bar/commit/abc123def456",
+                "CVE-2024-1234",
+            )
+            assert result is None
+
+
+# ---------------------------------------------------------------------------
+# OSVClient.get_fix_patch tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestOSVClientGetFixPatch:
+    """Test OSVClient.get_fix_patch flow."""
+
+    async def test_get_fix_patch_success(self):
+        """Successful flow: _query_osv returns data with FIX commit, fetch_from_url returns result."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        mock_fetcher = MagicMock(spec=WebPatchFetcher)
+        client = OSVClient(session=mock_session, patch_fetcher=mock_fetcher)
+
+        osv_data = {
+            "references": [
+                {"type": "FIX", "url": "https://github.com/openssl/openssl/commit/abc123"},
+            ],
+        }
+        mock_result = MagicMock(spec=WebPatchResult)
+        mock_result.repo_url = "https://github.com/openssl/openssl"
+
+        with patch.object(client, "_query_osv", new_callable=AsyncMock, return_value=osv_data):
+            mock_fetcher.fetch_from_url = AsyncMock(return_value=mock_result)
+            result = await client.get_fix_patch("CVE-2024-1234", "3.0.7", "openssl")
+
+        assert result is mock_result
+        # fetch_from_url should be called with the URL stripped of .patch suffix
+        mock_fetcher.fetch_from_url.assert_called_once_with(
+            "https://github.com/openssl/openssl/commit/abc123",
+            "CVE-2024-1234",
+            source="osv",
+        )
+
+    async def test_get_fix_patch_no_osv_data(self):
+        """When _query_osv returns None, get_fix_patch returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        mock_fetcher = MagicMock(spec=WebPatchFetcher)
+        client = OSVClient(session=mock_session, patch_fetcher=mock_fetcher)
+
+        with patch.object(client, "_query_osv", new_callable=AsyncMock, return_value=None):
+            result = await client.get_fix_patch("CVE-2024-0000")
+        assert result is None
+
+    async def test_get_fix_patch_no_fix_references(self):
+        """OSV data without FIX references returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        mock_fetcher = MagicMock(spec=WebPatchFetcher)
+        client = OSVClient(session=mock_session, patch_fetcher=mock_fetcher)
+
+        osv_data = {
+            "references": [
+                {"type": "ADVISORY", "url": "https://nvd.nist.gov/vuln/detail/CVE-2024-1234"},
+            ],
+        }
+
+        with patch.object(client, "_query_osv", new_callable=AsyncMock, return_value=osv_data):
+            result = await client.get_fix_patch("CVE-2024-1234")
+        assert result is None
+
+    async def test_get_fix_patch_exception_returns_none(self):
+        """Exception during flow returns None instead of propagating."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        mock_fetcher = MagicMock(spec=WebPatchFetcher)
+        client = OSVClient(session=mock_session, patch_fetcher=mock_fetcher)
+
+        with patch.object(client, "_query_osv", new_callable=AsyncMock, side_effect=RuntimeError("network error")):
+            result = await client.get_fix_patch("CVE-2024-1234")
+        assert result is None
+
+    async def test_get_fix_patch_strips_patch_suffix_for_github(self):
+        """GitHub .patch URLs should have the suffix stripped before resolution."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        mock_fetcher = MagicMock(spec=WebPatchFetcher)
+        client = OSVClient(session=mock_session, patch_fetcher=mock_fetcher)
+
+        osv_data = {
+            "references": [
+                {"type": "FIX", "url": "https://github.com/foo/bar/commit/abc123"},
+            ],
+        }
+        mock_result = MagicMock(spec=WebPatchResult)
+        mock_result.repo_url = "https://github.com/foo/bar"
+
+        with patch.object(client, "_query_osv", new_callable=AsyncMock, return_value=osv_data):
+            mock_fetcher.fetch_from_url = AsyncMock(return_value=mock_result)
+            result = await client.get_fix_patch("CVE-2024-1234")
+
+        # _extract_commit_from_references appends .patch, then get_fix_patch strips it
+        # for github URLs so that fetch_from_url can resolve them properly
+        call_url = mock_fetcher.fetch_from_url.call_args[0][0]
+        assert not call_url.endswith(".patch")
+
+
+# ---------------------------------------------------------------------------
+# _fetch_chromium_issue tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestFetchChromiumIssue:
+    """Test WebPatchFetcher._fetch_chromium_issue flow."""
+
+    async def test_fetch_chromium_issue_success(self):
+        """Full Chromium flow: issue URL -> Gerrit search -> CL selection -> patch fetch."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        from vuln_analysis.utils.gerrit_client import GerritChangeCandidate
+        candidate = GerritChangeCandidate(
+            submission_id=123, project="angle/angle", subject="Fix OOB read"
+        )
+
+        patch_text = (
+            "--- a/file.cc\n"
+            "+++ b/file.cc\n"
+            "@@ -1,1 +1,2 @@\n"
+            " line1\n"
+            "+line2\n"
+        )
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.search_changes_by_bug", new_callable=AsyncMock) as mock_search, \
+             patch("vuln_analysis.utils.web_patch_fetcher.list_merged_changes") as mock_list, \
+             patch("vuln_analysis.utils.web_patch_fetcher.select_gerrit_change", new_callable=AsyncMock) as mock_select, \
+             patch("vuln_analysis.utils.web_patch_fetcher.get_current_commit_sha", new_callable=AsyncMock) as mock_sha, \
+             patch("vuln_analysis.utils.web_patch_fetcher.project_to_gitiles_repo_url") as mock_repo_url, \
+             patch("vuln_analysis.utils.web_patch_fetcher.build_gitiles_patch_url") as mock_patch_url, \
+             patch.object(fetcher, "_fetch_patch_content", new_callable=AsyncMock) as mock_fetch:
+
+            mock_search.return_value = [{"_number": 123, "status": "MERGED"}]
+            mock_list.return_value = [candidate]
+            mock_select.return_value = 123
+            mock_sha.return_value = "abc123def456789abcdef0123456789abcdef012"
+            mock_repo_url.return_value = "https://chromium.googlesource.com/angle/angle"
+            mock_patch_url.return_value = "https://chromium.googlesource.com/angle/angle/+/abc123de%5E%21?format=TEXT"
+            mock_fetch.return_value = patch_text
+
+            result = await fetcher._fetch_chromium_issue(
+                "https://issues.chromium.org/issues/12345",
+                "CVE-2024-5678",
+                "ghsa",
+                "Buffer overflow in ANGLE",
+                None,
+            )
+            assert result is not None
+            assert result.platform == "gitiles"
+            assert result.url_type == "chromium_issue"
+            assert result.cve_id == "CVE-2024-5678"
+            assert result.source == "ghsa"
+
+    async def test_fetch_chromium_issue_invalid_url(self):
+        """Non-matching Chromium issue URL returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        result = await fetcher._fetch_chromium_issue(
+            "https://bugs.chromium.org/p/chromium/issues/detail?id=99999",
+            "CVE-2024-5678",
+            "ghsa",
+            None,
+            None,
+        )
+        assert result is None
+
+    async def test_fetch_chromium_issue_no_cls_found(self):
+        """Gerrit search returning empty results returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.search_changes_by_bug", new_callable=AsyncMock) as mock_search:
+            mock_search.return_value = []
+            result = await fetcher._fetch_chromium_issue(
+                "https://issues.chromium.org/issues/12345",
+                "CVE-2024-5678",
+                "ghsa",
+                None,
+                None,
+            )
+            assert result is None
+
+    async def test_fetch_chromium_issue_no_merged_cls(self):
+        """CLs found but none MERGED returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.search_changes_by_bug", new_callable=AsyncMock) as mock_search, \
+             patch("vuln_analysis.utils.web_patch_fetcher.list_merged_changes") as mock_list:
+            mock_search.return_value = [{"_number": 100, "status": "ABANDONED"}]
+            mock_list.return_value = []
+            result = await fetcher._fetch_chromium_issue(
+                "https://issues.chromium.org/issues/12345",
+                "CVE-2024-5678",
+                "ghsa",
+                None,
+                None,
+            )
+            assert result is None
+
+    async def test_fetch_chromium_issue_select_returns_none(self):
+        """When select_gerrit_change returns None, the flow returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        from vuln_analysis.utils.gerrit_client import GerritChangeCandidate
+        candidate = GerritChangeCandidate(
+            submission_id=100, project="chromium/src", subject="Fix thing"
+        )
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.search_changes_by_bug", new_callable=AsyncMock) as mock_search, \
+             patch("vuln_analysis.utils.web_patch_fetcher.list_merged_changes") as mock_list, \
+             patch("vuln_analysis.utils.web_patch_fetcher.select_gerrit_change", new_callable=AsyncMock) as mock_select:
+            mock_search.return_value = [{"_number": 100, "status": "MERGED"}]
+            mock_list.return_value = [candidate]
+            mock_select.return_value = None
+            result = await fetcher._fetch_chromium_issue(
+                "https://issues.chromium.org/issues/12345",
+                "CVE-2024-5678",
+                "ghsa",
+                None,
+                None,
+            )
+            assert result is None
+
+    async def test_fetch_chromium_issue_no_commit_sha(self):
+        """When get_current_commit_sha returns None, the flow returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        from vuln_analysis.utils.gerrit_client import GerritChangeCandidate
+        candidate = GerritChangeCandidate(
+            submission_id=100, project="chromium/src", subject="Fix thing"
+        )
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.search_changes_by_bug", new_callable=AsyncMock) as mock_search, \
+             patch("vuln_analysis.utils.web_patch_fetcher.list_merged_changes") as mock_list, \
+             patch("vuln_analysis.utils.web_patch_fetcher.select_gerrit_change", new_callable=AsyncMock) as mock_select, \
+             patch("vuln_analysis.utils.web_patch_fetcher.get_current_commit_sha", new_callable=AsyncMock) as mock_sha:
+            mock_search.return_value = [{"_number": 100, "status": "MERGED"}]
+            mock_list.return_value = [candidate]
+            mock_select.return_value = 100
+            mock_sha.return_value = None
+            result = await fetcher._fetch_chromium_issue(
+                "https://issues.chromium.org/issues/12345",
+                "CVE-2024-5678",
+                "ghsa",
+                None,
+                None,
+            )
+            assert result is None
+
+
+# ---------------------------------------------------------------------------
+# _fetch_patch_content + _decode_gitiles_response tests
+# ---------------------------------------------------------------------------
+
+@pytest.mark.asyncio
+class TestFetchPatchContent:
+    """Test WebPatchFetcher._fetch_patch_content."""
+
+    async def test_fetch_patch_content_standard_url(self):
+        """Non-Gitiles URL uses request_with_retry and returns text."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        patch_text = "--- a/file.c\n+++ b/file.c\n@@ -1 +1 @@\n-old\n+new\n"
+        mock_response = AsyncMock()
+        mock_response.text = AsyncMock(return_value=patch_text)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.request_with_retry") as mock_rwr:
+            mock_ctx = AsyncMock()
+            mock_ctx.__aenter__ = AsyncMock(return_value=mock_response)
+            mock_ctx.__aexit__ = AsyncMock(return_value=False)
+            mock_rwr.return_value = mock_ctx
+
+            result = await fetcher._fetch_patch_content("https://github.com/foo/bar/commit/abc123.patch")
+            assert result == patch_text
+
+    async def test_fetch_patch_content_404_returns_none(self):
+        """HTTP 404 returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.request_with_retry") as mock_rwr:
+            mock_rwr.side_effect = aiohttp.ClientResponseError(
+                request_info=MagicMock(),
+                history=(),
+                status=404,
+                message="Not Found",
+            )
+            result = await fetcher._fetch_patch_content("https://github.com/foo/bar/commit/deadbeef.patch")
+            assert result is None
+
+    async def test_fetch_patch_content_gitiles_url(self):
+        """Gitiles URL gets base64 decoded."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        raw_patch = "--- a/file.cc\n+++ b/file.cc\n"
+        b64_content = base64.b64encode(raw_patch.encode()).decode()
+
+        with patch.object(fetcher, "_fetch_gitiles_patch", new_callable=AsyncMock, return_value=b64_content):
+            result = await fetcher._fetch_patch_content(
+                "https://chromium.googlesource.com/angle/angle/+/abc123%5E%21?format=TEXT"
+            )
+            assert result == raw_patch
+
+    async def test_fetch_patch_content_gitiles_fetch_fails(self):
+        """Gitiles URL where _fetch_gitiles_patch returns None returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch.object(fetcher, "_fetch_gitiles_patch", new_callable=AsyncMock, return_value=None):
+            result = await fetcher._fetch_patch_content(
+                "https://chromium.googlesource.com/angle/angle/+/abc123%5E%21?format=TEXT"
+            )
+            assert result is None
+
+    async def test_fetch_patch_content_rate_limited_returns_none(self):
+        """HTTP 429 returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.request_with_retry") as mock_rwr:
+            mock_rwr.side_effect = aiohttp.ClientResponseError(
+                request_info=MagicMock(),
+                history=(),
+                status=429,
+                message="Too Many Requests",
+            )
+            result = await fetcher._fetch_patch_content("https://github.com/foo/bar/commit/abc123.patch")
+            assert result is None
+
+    async def test_fetch_patch_content_generic_exception_returns_none(self):
+        """Generic exception during fetch returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+
+        with patch("vuln_analysis.utils.web_patch_fetcher.request_with_retry") as mock_rwr:
+            mock_rwr.side_effect = ConnectionResetError("Connection reset by peer")
+            result = await fetcher._fetch_patch_content("https://github.com/foo/bar/commit/abc123.patch")
+            assert result is None
+
+
+class TestDecodeGitilesResponse:
+    """Test WebPatchFetcher._decode_gitiles_response."""
+
+    def test_decode_gitiles_response_valid(self):
+        """Valid base64-encoded content is decoded correctly."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+        raw = "--- a/file.c\n+++ b/file.c\n"
+        encoded = base64.b64encode(raw.encode()).decode()
+        result = fetcher._decode_gitiles_response(encoded)
+        assert result == raw
+
+    def test_decode_gitiles_response_invalid(self):
+        """Invalid base64 content returns None."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+        result = fetcher._decode_gitiles_response("not-valid-base64!!!")
+        assert result is None
+
+    def test_decode_gitiles_response_empty_string(self):
+        """Empty string base64-decoded is empty string."""
+        mock_session = MagicMock(spec=aiohttp.ClientSession)
+        fetcher = WebPatchFetcher(session=mock_session)
+        encoded = base64.b64encode(b"").decode()
+        result = fetcher._decode_gitiles_response(encoded)
+        assert result == ""
